reverting changes - not working out.. rethinking approach
