+++ app.py
@@ -5,18 +5,18 @@ import re
 import json
 import threading
 import time
-import queue # For log queue
+import queue # Added for log queue
 from datetime import datetime, timedelta, timezone
 import random
 
 import docker
 from docker.errors import NotFound, APIError
-# Added Response, stream_with_context, render_template_string
-from flask import Flask, jsonify, render_template, redirect, url_for, request, Response, stream_with_context, render_template_string
+from flask import Flask, jsonify, render_template, redirect, url_for, request, Response, stream_with_context # Added Response, stream_with_context
 from dotenv import load_dotenv
 import requests
 
 # --- Configuration ---
+# Basic logging config will be augmented later for queue
 logging.basicConfig(level=logging.INFO, format='%(asctime)s [%(levelname)s] [%(threadName)s] %(message)s')
 load_dotenv()
 
@@ -66,26 +66,20 @@ class QueueLogHandler(logging.Handler):
         self.log_queue = log_queue
 
     def emit(self, record):
-        # Check if the message is already formatted for SSE (our custom events)
-        # This prevents double-formatting log messages if they somehow contain "event:"
         log_entry = self.format(record)
-        if "event:" not in log_entry and "data:" not in log_entry:
-            # Format as a standard SSE message event if it's a normal log record
-             message_payload = json.dumps(log_entry) # Ensure valid JSON payload
-             sse_message = f"event: message\ndata: {message_payload}\n\n"
-        else:
-             # Assume it's a pre-formatted SSE event (like rules_update)
-             sse_message = log_entry # Pass it through directly
-
         try:
-            self.log_queue.put_nowait(sse_message)
+            self.log_queue.put_nowait(log_entry)
         except queue.Full:
+            # Try removing oldest item to make space
             try:
-                self.log_queue.get_nowait() # Discard oldest item
-                self.log_queue.put_nowait(sse_message) # Try putting again
-            except queue.Empty: pass
-            except queue.Full: print(f"Log queue full, dropping message: {log_entry[:100]}...", file=sys.stderr)
+                self.log_queue.get_nowait()
+                self.log_queue.put_nowait(log_entry) # Try putting again
+            except queue.Empty:
+                pass # Queue was actually empty, weird race condition?
+            except queue.Full:
+                 print("Log queue full, dropping message.", file=sys.stderr) # Log to stderr directly
 
+# Add Queue Handler to root logger
 queue_handler = QueueLogHandler(log_queue)
 queue_handler.setFormatter(log_formatter)
 queue_handler.setLevel(logging.INFO) # Only push INFO and above to the web UI
@@ -94,687 +88,1729 @@ root_logger.addHandler(queue_handler)
 
 # Docker Client Setup
 try:
-    docker_client = docker.from_env(timeout=10); docker_client.ping()
+    docker_client = docker.from_env(timeout=10)
+    docker_client.ping()
     logging.info("Successfully connected to Docker daemon.")
-except Exception as e: logging.error(f"FATAL: Failed to connect to Docker daemon: {e}"); docker_client = None
+except Exception as e:
+    logging.error(f"FATAL: Failed to connect to Docker daemon: {e}")
+    docker_client = None
 
 # Global State
 tunnel_state = { "name": TUNNEL_NAME, "id": None, "token": None, "status_message": "Initializing...", "error": None }
 cloudflared_agent_state = { "container_status": "unknown", "last_action_status": None }
-managed_rules = {}; zone_id_cache = {}; state_lock = threading.Lock(); stop_event = threading.Event()
-
-# --- NEW: Helper to Send SSE Events ---
-def send_sse_event(event_type, data=""):
-    """Puts a specially formatted message into the log queue for SSE."""
-    sse_message = f"event: {event_type}\ndata: {json.dumps(data)}\n\n"
-    try: log_queue.put_nowait(sse_message); logging.debug(f"Sent SSE event: {event_type}")
-    except queue.Full: logging.warning(f"Log queue full when trying to send SSE event: {event_type}. Event lost.")
+managed_rules = {} # Stores hostname -> {service, container_id, status, delete_at, zone_id}
+zone_id_cache = {} # Stores zone_name -> zone_id
+state_lock = threading.Lock()
+stop_event = threading.Event()
 
 
 def load_state():
-    global managed_rules; state_dir = os.path.dirname(STATE_FILE_PATH)
+    global managed_rules
+    state_dir = os.path.dirname(STATE_FILE_PATH)
     if not os.path.exists(state_dir):
-        try: os.makedirs(state_dir, exist_ok=True); logging.info(f"Created directory for state file: {state_dir}")
-        except OSError as e: logging.error(f"FATAL: Could not create directory for state file {state_dir}: {e}. State persistence will fail."); managed_rules = {}; return
-    if not os.path.exists(STATE_FILE_PATH): logging.info(f"State file '{STATE_FILE_PATH}' not found, starting fresh."); managed_rules = {}; return
+        try:
+             os.makedirs(state_dir, exist_ok=True)
+             logging.info(f"Created directory for state file: {state_dir}")
+        except OSError as e:
+             logging.error(f"FATAL: Could not create directory for state file {state_dir}: {e}. State persistence will fail.")
+             managed_rules = {}
+             return
+
+    if not os.path.exists(STATE_FILE_PATH):
+        logging.info(f"State file '{STATE_FILE_PATH}' not found, starting fresh.")
+        managed_rules = {}
+        return
     try:
-        with open(STATE_FILE_PATH, 'r') as f: loaded_data = json.load(f)
+        with open(STATE_FILE_PATH, 'r') as f:
+            loaded_data = json.load(f)
         for hostname, rule in loaded_data.items():
+             # Ensure delete_at is converted back to datetime object
              if rule.get("delete_at") and isinstance(rule.get("delete_at"), str):
                  try:
-                     if rule["delete_at"].endswith('Z'): dt = datetime.fromisoformat(rule["delete_at"].replace('Z', '+00:00'))
-                     else: dt = datetime.fromisoformat(rule["delete_at"])
+                     if rule["delete_at"].endswith('Z'):
+                        dt = datetime.fromisoformat(rule["delete_at"].replace('Z', '+00:00'))
+                     else:
+                         dt = datetime.fromisoformat(rule["delete_at"])
+                     # Ensure timezone aware (assume UTC if naive)
                      rule["delete_at"] = dt.replace(tzinfo=timezone.utc) if dt.tzinfo is None else dt.astimezone(timezone.utc)
-                 except ValueError as date_err: logging.warning(f"Could not parse delete_at for {hostname}: {rule['delete_at']} Error: {date_err}. Setting to None."); rule["delete_at"] = None
-             elif not isinstance(rule.get("delete_at"), datetime): rule["delete_at"] = None
-             if "zone_id" not in rule: logging.warning(f"Rule for {hostname} loaded from state missing 'zone_id'. Will re-determine."); rule["zone_id"] = None
-        managed_rules = loaded_data; logging.info(f"Loaded state for {len(managed_rules)} rules from {STATE_FILE_PATH}")
-    except (json.JSONDecodeError, IOError, OSError) as e: logging.error(f"Error loading state from {STATE_FILE_PATH}: {e}. Starting fresh.", exc_info=True); managed_rules = {}
+                 except ValueError as date_err:
+                     logging.warning(f"Could not parse delete_at for {hostname}: {rule['delete_at']} Error: {date_err}. Setting to None.")
+                     rule["delete_at"] = None
+             elif not isinstance(rule.get("delete_at"), datetime):
+                 rule["delete_at"] = None
+             # Ensure zone_id field exists
+             if "zone_id" not in rule:
+                 logging.warning(f"Rule for {hostname} loaded from state is missing 'zone_id'. Will attempt to re-determine on reconcile.")
+                 rule["zone_id"] = None # Mark as unknown for now
+        managed_rules = loaded_data
+        logging.info(f"Loaded state for {len(managed_rules)} rules from {STATE_FILE_PATH}")
+    except (json.JSONDecodeError, IOError, OSError) as e:
+        logging.error(f"Error loading state from {STATE_FILE_PATH}: {e}. Starting fresh.", exc_info=True)
+        managed_rules = {}
+
 
 def save_state():
     serializable_state = {}
     for hostname, rule in managed_rules.items():
         rule_copy = rule.copy()
-        if rule_copy.get("delete_at") and isinstance(rule_copy["delete_at"], datetime): rule_copy["delete_at"] = rule_copy["delete_at"].astimezone(timezone.utc).strftime('%Y-%m-%dT%H:%M:%SZ')
-        if "zone_id" not in rule_copy: logging.warning(f"Attempting to save rule for {hostname} without zone_id!"); rule_copy["zone_id"] = None
+        if rule_copy.get("delete_at") and isinstance(rule_copy["delete_at"], datetime):
+            rule_copy["delete_at"] = rule_copy["delete_at"].astimezone(timezone.utc).strftime('%Y-%m-%dT%H:%M:%SZ')
+        if "zone_id" not in rule_copy:
+            logging.warning(f"Attempting to save rule for {hostname} without zone_id!")
+            rule_copy["zone_id"] = None
         serializable_state[hostname] = rule_copy
     try:
         state_dir = os.path.dirname(STATE_FILE_PATH)
         if not os.path.exists(state_dir):
             try: os.makedirs(state_dir, exist_ok=True); logging.info(f"Created directory {state_dir} before saving state.")
             except OSError as e: logging.error(f"Could not create directory {state_dir} for state file: {e}. Save failed."); return
+
         temp_file_path = STATE_FILE_PATH + ".tmp"
-        with open(temp_file_path, 'w') as f: json.dump(serializable_state, f, indent=2)
-        os.replace(temp_file_path, STATE_FILE_PATH); logging.debug(f"Saved state for {len(managed_rules)} rules to {STATE_FILE_PATH}")
-    except (IOError, OSError) as e: logging.error(f"Error saving state to {STATE_FILE_PATH}: {e}", exc_info=True)
+        with open(temp_file_path, 'w') as f:
+            json.dump(serializable_state, f, indent=2)
+        os.replace(temp_file_path, STATE_FILE_PATH)
+        logging.debug(f"Saved state for {len(managed_rules)} rules to {STATE_FILE_PATH}")
+    except (IOError, OSError) as e:
+        logging.error(f"Error saving state to {STATE_FILE_PATH}: {e}", exc_info=True)
+
 
 def cf_api_request(method, endpoint, json_data=None, params=None):
-    url = f"{CF_API_BASE_URL}{endpoint}"; error_msg = None
+    url = f"{CF_API_BASE_URL}{endpoint}"
+    error_msg = None
     try:
-        logging.info(f"API Request: {method} {url} Params: {params}")
-        if json_data: logging.debug(f"API Request Data: {json_data}")
+        logging.info(f"API Request: {method} {url} Params: {params}") # Don't log full data by default
+        if json_data:
+             logging.debug(f"API Request Data: {json_data}")
         response = requests.request(method, url, headers=CF_HEADERS, json=json_data, params=params, timeout=30)
-        response.raise_for_status(); logging.info(f"API Response Status: {response.status_code}")
-        if response.status_code == 204 or not response.content: return {"success": True, "result": None}
+        response.raise_for_status()
+        logging.info(f"API Response Status: {response.status_code}")
+        if response.status_code == 204 or not response.content:
+            return {"success": True, "result": None}
         try:
-            response_data = response.json(); logging.debug(f"API Response Body (first 500 chars): {str(response_data)[:500]}")
+            response_data = response.json()
+            logging.debug(f"API Response Body (first 500 chars): {str(response_data)[:500]}")
             if isinstance(response_data, dict) and 'success' in response_data:
-                 if response_data['success']: return response_data
+                 if response_data['success']:
+                      return response_data
                  else:
                       cf_errors = response_data.get('errors', [])
-                      if cf_errors and isinstance(cf_errors, list) and len(cf_errors) > 0 and isinstance(cf_errors[0], dict): error_msg = f"API Error: {cf_errors[0].get('message', 'Unknown error')}"
-                      else: error_msg = f"API reported failure but no error details provided. Response: {response_data}"
-                      logging.error(f"API Request Failed ({method} {url}): {error_msg} - Full Errors: {cf_errors}"); raise requests.exceptions.RequestException(error_msg, response=response)
-            else: logging.warning(f"API response for {method} {url} missing 'success' field. Status: {response.status_code}. Body: {str(response_data)[:200]}"); raise requests.exceptions.RequestException(f"Unexpected JSON response format from API. Status: {response.status_code}", response=response)
-        except json.JSONDecodeError: logging.error(f"API response for {method} {url} not valid JSON. Status: {response.status_code}. Body: {response.text[:200]}"); raise requests.exceptions.RequestException(f"Invalid JSON response from API. Status: {response.status_code}", response=response)
+                      if cf_errors and isinstance(cf_errors, list) and len(cf_errors) > 0 and isinstance(cf_errors[0], dict):
+                           error_msg = f"API Error: {cf_errors[0].get('message', 'Unknown error')}"
+                      else:
+                           error_msg = f"API reported failure but no error details provided. Response: {response_data}"
+                      logging.error(f"API Request Failed ({method} {url}): {error_msg} - Full Errors: {cf_errors}")
+                      raise requests.exceptions.RequestException(error_msg, response=response)
+            else:
+                 logging.warning(f"API response for {method} {url} was valid JSON but missing 'success' field. Status: {response.status_code}. Body: {str(response_data)[:200]}")
+                 raise requests.exceptions.RequestException(f"Unexpected JSON response format from API. Status: {response.status_code}", response=response)
+        except json.JSONDecodeError:
+            logging.error(f"API response for {method} {url} was not valid JSON. Status: {response.status_code}. Body: {response.text[:200]}")
+            raise requests.exceptions.RequestException(f"Invalid JSON response from API. Status: {response.status_code}", response=response)
     except requests.exceptions.RequestException as e:
         if error_msg is None:
-            logging.error(f"API Request Failed: {method} {url}"); error_msg = f"Request Exception: {e}"
+            logging.error(f"API Request Failed: {method} {url}")
+            error_msg = f"Request Exception: {e}"
             if e.response is not None:
                 try:
-                    error_data = e.response.json(); cf_errors = error_data.get('errors', [])
-                    if cf_errors and isinstance(cf_errors, list) and len(cf_errors) > 0 and isinstance(cf_errors[0], dict): error_msg = f"API Error: {cf_errors[0].get('message', 'Unknown error')}"
-                    else: error_msg = f"HTTP {e.response.status_code} - {e.response.text[:100]}"
+                    error_data = e.response.json()
+                    cf_errors = error_data.get('errors', [])
+                    if cf_errors and isinstance(cf_errors, list) and len(cf_errors) > 0 and isinstance(cf_errors[0], dict):
+                        error_msg = f"API Error: {cf_errors[0].get('message', 'Unknown error')}"
+                    else:
+                         error_msg = f"HTTP {e.response.status_code} - {e.response.text[:100]}"
                     logging.error(f"API Error Response Body: {error_data}")
-                except (ValueError, AttributeError, json.JSONDecodeError): error_msg = f"HTTP {e.response.status_code} - {e.response.text[:100]}"
+                except (ValueError, AttributeError, json.JSONDecodeError):
+                     error_msg = f"HTTP {e.response.status_code} - {e.response.text[:100]}"
             logging.error(f"Final error message: {error_msg}")
-        if "cfd_tunnel" in endpoint and tunnel_state.get("id") is None and "token" not in endpoint: tunnel_state["error"] = error_msg
+
+        # Capture initial tunnel setup errors
+        if "cfd_tunnel" in endpoint and tunnel_state.get("id") is None and "token" not in endpoint:
+             tunnel_state["error"] = error_msg
         raise requests.exceptions.RequestException(error_msg, response=e.response)
 
+
 def get_zone_id_from_name(zone_name):
+    """Retrieves the Zone ID for a given zone name, using cache."""
     global zone_id_cache
-    if not zone_name: logging.warning("get_zone_id_from_name called with empty zone_name."); return None
-    with state_lock: cached_id = zone_id_cache.get(zone_name)
-    if cached_id: logging.debug(f"Zone ID for '{zone_name}' found in cache: {cached_id}"); return cached_id
+    if not zone_name:
+        logging.warning("get_zone_id_from_name called with empty zone_name.")
+        return None
+
+    with state_lock:
+        cached_id = zone_id_cache.get(zone_name)
+    if cached_id:
+        logging.debug(f"Zone ID for '{zone_name}' found in cache: {cached_id}")
+        return cached_id
+
     logging.info(f"Zone ID for '{zone_name}' not in cache. Querying Cloudflare API...")
-    endpoint = "/zones"; params = {"name": zone_name, "status": "active"}
+    endpoint = "/zones"
+    params = {"name": zone_name, "status": "active"}
+
     try:
-        response_data = cf_api_request("GET", endpoint, params=params); results = response_data.get("result", [])
+        response_data = cf_api_request("GET", endpoint, params=params)
+        results = response_data.get("result", [])
+
         if results and isinstance(results, list) and len(results) == 1:
-            zone_id = results[0].get("id"); zone_actual_name = results[0].get("name")
-            # *** CORRECTED BLOCK ***
+            zone_id = results[0].get("id")
+            zone_actual_name = results[0].get("name")
             if zone_id and zone_actual_name == zone_name:
                 logging.info(f"Found Zone ID for '{zone_name}': {zone_id}")
-                # Acquire lock ONLY to update the shared cache
                 with state_lock:
                     zone_id_cache[zone_name] = zone_id
-                # Return the zone_id *after* the lock is released
                 return zone_id
-            # *** END CORRECTED BLOCK ***
-            else: logging.error(f"API returned unexpected result or name/ID mismatch for zone '{zone_name}': {results[0]}"); return None
-        elif results and len(results) > 1: logging.error(f"API returned multiple ({len(results)}) active zones matching name '{zone_name}'. Cannot determine correct zone."); return None
-        else: logging.warning(f"No active zone found matching name '{zone_name}' via API."); return None
-    except requests.exceptions.RequestException as e: logging.error(f"API error looking up zone '{zone_name}': {e}"); return None
-    except Exception as e: logging.error(f"Unexpected error looking up zone '{zone_name}': {e}", exc_info=True); return None
+            else:
+                logging.error(f"API returned unexpected result or name mismatch for zone '{zone_name}': {results[0]}")
+                return None
+        elif results and len(results) > 1:
+            logging.error(f"API returned multiple ({len(results)}) active zones matching name '{zone_name}'. Cannot determine correct zone.")
+            return None
+        else:
+            logging.warning(f"No active zone found matching name '{zone_name}' via API.")
+            return None
+    except requests.exceptions.RequestException as e:
+        logging.error(f"API error looking up zone '{zone_name}': {e}")
+        return None
+    except Exception as e:
+        logging.error(f"Unexpected error looking up zone '{zone_name}': {e}", exc_info=True)
+        return None
+
 
 def find_tunnel_via_api(name):
-    logging.info(f"Finding tunnel '{name}' via API"); endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel"; params = {"name": name, "is_deleted": "false"}
+    """Finds an existing tunnel and its token via the API."""
+    logging.info(f"Finding tunnel '{name}' via API")
+    endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel"
+    params = {"name": name, "is_deleted": "false"}
     try:
-        response_data = cf_api_request("GET", endpoint, params=params); tunnels = response_data.get("result", [])
+        response_data = cf_api_request("GET", endpoint, params=params)
+        tunnels = response_data.get("result", [])
         if tunnels and isinstance(tunnels, list):
-            tunnel = tunnels[0]; tunnel_id = tunnel.get("id")
-            if tunnel_id: logging.info(f"Found existing tunnel '{name}' ID: {tunnel_id}. Getting token..."); token = get_tunnel_token_via_api(tunnel_id); return tunnel_id, token
-            else: logging.warning(f"Found tunnel entry for '{name}' but it has no ID: {tunnel}"); return None, None
-        else: logging.info(f"Tunnel '{name}' not found via API."); return None, None
-    except requests.exceptions.RequestException as e: logging.error(f"API error finding tunnel '{name}': {e}"); return None, None
-    except Exception as e: logging.error(f"Unexpected error finding tunnel '{name}': {e}", exc_info=True); tunnel_state["error"] = f"Unexpected error finding tunnel: {e}"; return None, None
+            tunnel = tunnels[0]
+            tunnel_id = tunnel.get("id")
+            if tunnel_id:
+                logging.info(f"Found existing tunnel '{name}' ID: {tunnel_id}. Getting token...")
+                token = get_tunnel_token_via_api(tunnel_id)
+                return tunnel_id, token
+            else:
+                 logging.warning(f"Found tunnel entry for '{name}' but it has no ID: {tunnel}")
+                 return None, None
+        else:
+            logging.info(f"Tunnel '{name}' not found via API.")
+            return None, None
+    except requests.exceptions.RequestException as e:
+        logging.error(f"API error finding tunnel '{name}': {e}")
+        return None, None
+    except Exception as e:
+        logging.error(f"Unexpected error finding tunnel '{name}': {e}", exc_info=True)
+        tunnel_state["error"] = f"Unexpected error finding tunnel: {e}"
+        return None, None
+
 
 def get_tunnel_token_via_api(tunnel_id):
-    logging.info(f"Getting token for tunnel ID '{tunnel_id}'"); endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel/{tunnel_id}/token"; url = f"{CF_API_BASE_URL}{endpoint}"
+    """Gets the token for a specific tunnel ID."""
+    logging.info(f"Getting token for tunnel ID '{tunnel_id}'")
+    endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel/{tunnel_id}/token"
+    url = f"{CF_API_BASE_URL}{endpoint}"
     try:
+        # Note: Using requests directly here because cf_api_request expects JSON response
         logging.info(f"API Request: GET {url} (for token)")
-        response = requests.request("GET", url, headers={"Authorization": f"Bearer {CF_API_TOKEN}"}, timeout=30); response.raise_for_status()
+        response = requests.request("GET", url, headers={"Authorization": f"Bearer {CF_API_TOKEN}"}, timeout=30)
+        response.raise_for_status()
         token = response.text.strip()
-        if not token or len(token) < 50: logging.error(f"Retrieved token for tunnel {tunnel_id} appears invalid."); raise ValueError("Invalid token format received from API")
-        logging.info(f"Successfully retrieved token for tunnel {tunnel_id}"); return token
+        if not token or len(token) < 50: # Basic sanity check on token format
+            logging.error(f"Retrieved token for tunnel {tunnel_id} appears invalid.")
+            raise ValueError("Invalid token format received from API")
+        logging.info(f"Successfully retrieved token for tunnel {tunnel_id}")
+        return token
     except requests.exceptions.RequestException as e:
-        error_msg = f"API Error getting token for tunnel {tunnel_id}: {e}";
-        if e.response is not None: error_msg += f" Status: {e.response.status_code} Body: {e.response.text[:100]}"
-        logging.error(error_msg); tunnel_state["error"] = error_msg; raise
-    except Exception as e: logging.error(f"Unexpected error getting tunnel token for {tunnel_id}: {e}", exc_info=True); tunnel_state["error"] = f"Unexpected error getting token: {e}"; raise
+        error_msg = f"API Error getting token for tunnel {tunnel_id}: {e}"
+        if e.response is not None:
+             error_msg += f" Status: {e.response.status_code} Body: {e.response.text[:100]}"
+        logging.error(error_msg)
+        tunnel_state["error"] = error_msg
+        raise # Re-raise to be caught by caller
+    except Exception as e:
+         logging.error(f"Unexpected error getting tunnel token for {tunnel_id}: {e}", exc_info=True)
+         tunnel_state["error"] = f"Unexpected error getting token: {e}"
+         raise
+
 
 def create_tunnel_via_api(name):
-    logging.info(f"Creating tunnel '{name}' via API"); endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel"; payload = {"name": name, "config_src": "cloudflare"}
+    """Creates a new tunnel and returns its ID and token."""
+    logging.info(f"Creating tunnel '{name}' via API")
+    endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel"
+    payload = {"name": name, "config_src": "cloudflare"}
     try:
-        response_data = cf_api_request("POST", endpoint, json_data=payload); result = response_data.get("result", {}); tunnel_id = result.get("id"); token = result.get("token")
-        if not tunnel_id or not token: logging.error(f"API response for tunnel creation missing ID or Token: {result}"); raise ValueError("Missing ID or Token in API response")
-        logging.info(f"Successfully created tunnel '{name}' with ID {tunnel_id}."); return tunnel_id, token
-    except requests.exceptions.RequestException as e: logging.error(f"API error creating tunnel '{name}': {e}"); return None, None
-    except Exception as e: logging.error(f"Unexpected error creating tunnel '{name}': {e}", exc_info=True); tunnel_state["error"] = f"Unexpected error creating tunnel: {e}"; return None, None
+        response_data = cf_api_request("POST", endpoint, json_data=payload)
+        result = response_data.get("result", {})
+        tunnel_id = result.get("id")
+        token = result.get("token")
+        if not tunnel_id or not token:
+            logging.error(f"API response for tunnel creation missing ID or Token: {result}")
+            raise ValueError("Missing ID or Token in API response")
+        logging.info(f"Successfully created tunnel '{name}' with ID {tunnel_id}.")
+        return tunnel_id, token
+    except requests.exceptions.RequestException as e:
+        logging.error(f"API error creating tunnel '{name}': {e}")
+        return None, None # Return None on failure
+    except Exception as e:
+        logging.error(f"Unexpected error creating tunnel '{name}': {e}", exc_info=True)
+        tunnel_state["error"] = f"Unexpected error creating tunnel: {e}"
+        return None, None
+
 
 def initialize_tunnel():
-    logging.info("Initializing tunnel..."); tunnel_state["status_message"] = f"Checking for tunnel '{TUNNEL_NAME}' via API..."; tunnel_state["error"] = None
-    tunnel_id = None; token = None
+    """Finds or creates the tunnel and gets its token."""
+    logging.info("Initializing tunnel...")
+    tunnel_state["status_message"] = f"Checking for tunnel '{TUNNEL_NAME}' via API..."
+    tunnel_state["error"] = None
+    tunnel_id = None
+    token = None
     try:
         tunnel_id, token = find_tunnel_via_api(TUNNEL_NAME)
-        if not tunnel_id and not tunnel_state.get("error"): tunnel_state["status_message"] = f"Tunnel '{TUNNEL_NAME}' not found. Creating via API..."; tunnel_id, token = create_tunnel_via_api(TUNNEL_NAME)
-        if tunnel_id and token: tunnel_state["id"] = tunnel_id; tunnel_state["token"] = token; tunnel_state["status_message"] = "Tunnel setup complete (using API)."; tunnel_state["error"] = None; logging.info(f"Tunnel '{TUNNEL_NAME}' initialized successfully. ID: {tunnel_id}")
-        elif not tunnel_state.get("error"): tunnel_state["status_message"] = "Tunnel initialization failed."; tunnel_state["error"] = "Failed to find/create tunnel or retrieve token."; logging.error(f"Tunnel initialization failed for '{TUNNEL_NAME}'.")
-        else: tunnel_state["status_message"] = "Tunnel initialization failed (see error details)."
+
+        if not tunnel_id and not tunnel_state.get("error"): # Only create if find didn't fail
+            tunnel_state["status_message"] = f"Tunnel '{TUNNEL_NAME}' not found. Creating via API..."
+            tunnel_id, token = create_tunnel_via_api(TUNNEL_NAME)
+
+        if tunnel_id and token:
+            tunnel_state["id"] = tunnel_id
+            tunnel_state["token"] = token
+            tunnel_state["status_message"] = "Tunnel setup complete (using API)."
+            tunnel_state["error"] = None
+            logging.info(f"Tunnel '{TUNNEL_NAME}' initialized successfully. ID: {tunnel_id}")
+        elif not tunnel_state.get("error"):
+             # If we reached here without ID/token and no prior error, something went wrong
+             tunnel_state["status_message"] = "Tunnel initialization failed."
+             tunnel_state["error"] = "Failed to find/create tunnel or retrieve token. Check logs."
+             logging.error(f"Tunnel initialization failed for '{TUNNEL_NAME}'. Could not get ID and Token.")
+        else:
+             # An error occurred during find/create/token retrieval
+             tunnel_state["status_message"] = "Tunnel initialization failed (see error details)."
+             # Error message should already be in tunnel_state["error"] from the failing function
+
         logging.info(f"Tunnel init completed. State: ID={tunnel_state.get('id')}, Token Present={bool(tunnel_state.get('token'))}, Error={tunnel_state.get('error')}")
-    except Exception as e: logging.error(f"Unhandled exception during tunnel initialization: {e}", exc_info=True); if not tunnel_state.get("error"): tunnel_state["error"] = f"Initialization failed unexpectedly: {e}"; tunnel_state["status_message"] = "Tunnel initialization failed (unexpected error)."
+
+    except Exception as e:
+        # Catch unexpected errors during the init flow
+        logging.error(f"Unhandled exception during tunnel initialization: {e}", exc_info=True)
+        if not tunnel_state.get("error"): # Avoid overwriting specific API errors
+            tunnel_state["error"] = f"Initialization failed unexpectedly: {e}"
+        tunnel_state["status_message"] = "Tunnel initialization failed (unexpected error)."
+
 
 def get_current_cf_config():
-    if not tunnel_state.get("id"): logging.warning("Cannot get CF config, tunnel ID not available."); return None
+    """Gets the current tunnel configuration from Cloudflare."""
+    if not tunnel_state.get("id"):
+        logging.warning("Cannot get CF config, tunnel ID not available.")
+        return None
+
     logging.debug("Fetching current CF tunnel configuration.")
     endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel/{tunnel_state['id']}/configurations"
     try:
         response_data = cf_api_request("GET", endpoint)
         if response_data and response_data.get("success"):
-            result_data = response_data.get("result"); config_data = None
-            if isinstance(result_data, dict): config_data = result_data.get("config")
-            if isinstance(config_data, dict): logging.debug(f"Fetched config: {config_data}"); return config_data
-            elif config_data is None: logging.info("Fetched config is null."); return {}
-            else: logging.warning(f"Unexpected type for 'config': {type(config_data)}."); return {}
-        else: logging.error(f"Get config API call failed: {response_data}"); return None
+            result_data = response_data.get("result")
+            config_data = None
+            if isinstance(result_data, dict):
+                 config_data = result_data.get("config")
+
+            if isinstance(config_data, dict):
+                 logging.debug(f"Fetched config: {config_data}")
+                 return config_data
+            elif config_data is None:
+                 # API returns null config if never configured
+                 logging.info("Fetched config is null.")
+                 return {} # Return empty dict for consistency
+            else:
+                 logging.warning(f"Unexpected type for 'config' field in API response: {type(config_data)}. Result: {result_data}")
+                 return {} # Return empty dict
+        else:
+            logging.error(f"Get config API call failed or returned success=false: {response_data}")
+            return None
     except requests.exceptions.RequestException as e:
+        # *** CORRECTED BLOCK ***
         logging.error(f"API error fetching config: {e}")
-        if not tunnel_state.get("error"): tunnel_state["error"] = f"Failed get tunnel config: {e}"
-        return None
+        # Only set the global error if one isn't already set (avoid overwriting more specific errors)
+        if not tunnel_state.get("error"):
+            tunnel_state["error"] = f"Failed get tunnel config: {e}"
+        return None # Return None as the config couldn't be fetched
+        # *** END CORRECTED BLOCK ***
     except Exception as e:
         logging.error(f"Unexpected error fetching config: {e}", exc_info=True)
-        if not tunnel_state.get("error"): tunnel_state["error"] = f"Unexpected error getting tunnel config: {e}"
+        if not tunnel_state.get("error"):
+            tunnel_state["error"] = f"Unexpected error getting tunnel config: {e}"
         return None
 
+
 def find_dns_record_id(zone_id, hostname, tunnel_id):
-    if not zone_id or not hostname or not tunnel_id: logging.error("find_dns_record_id: Missing args."); return None
-    expected_content = f"{tunnel_id}.cfargotunnel.com"; endpoint = f"/zones/{zone_id}/dns_records"; params = {"type": "CNAME", "name": hostname, "content": expected_content, "match": "all"}
+    """Finds the ID of a specific CNAME DNS record pointing to the tunnel."""
+    if not zone_id or not hostname or not tunnel_id:
+        logging.error("find_dns_record_id: Missing required arguments.")
+        return None
+
+    expected_content = f"{tunnel_id}.cfargotunnel.com"
+    endpoint = f"/zones/{zone_id}/dns_records"
+    params = {"type": "CNAME", "name": hostname, "content": expected_content, "match": "all"}
+
     try:
         logging.info(f"Searching DNS: Zone={zone_id}, Type=CNAME, Name={hostname}, Content={expected_content}")
-        response_data = cf_api_request("GET", endpoint, params=params); results = response_data.get("result", [])
+        response_data = cf_api_request("GET", endpoint, params=params)
+        results = response_data.get("result", [])
+
         if results and isinstance(results, list):
+            # API might return multiple if pagination is involved, but match=all should limit it
             record_id = results[0].get("id")
-            if record_id: logging.info(f"Found DNS record for {hostname} in zone {zone_id} with ID: {record_id}"); return record_id
-            else: logging.warning(f"DNS record found for {hostname} but lacks ID: {results[0]}"); return None
-        else: logging.info(f"No matching DNS record found for {hostname} in zone {zone_id}"); return None
-    except requests.exceptions.RequestException as e: logging.error(f"API error finding DNS record for {hostname}: {e}"); return None
-    except Exception as e: logging.error(f"Unexpected error finding DNS record for {hostname}: {e}", exc_info=True); return None
+            if record_id:
+                logging.info(f"Found DNS record for {hostname} in zone {zone_id} with ID: {record_id}")
+                return record_id
+            else:
+                logging.warning(f"DNS record found for {hostname} but it lacks an ID field: {results[0]}")
+                return None
+        else:
+            logging.info(f"No matching DNS record found for {hostname} in zone {zone_id}")
+            return None
+    except requests.exceptions.RequestException as e:
+        logging.error(f"API error finding DNS record for {hostname}: {e}")
+        return None
+    except Exception as e:
+        logging.error(f"Unexpected error finding DNS record for {hostname}: {e}", exc_info=True)
+        return None
+
 
 def create_cloudflare_dns_record(zone_id, hostname, tunnel_id):
-    if not zone_id or not hostname or not tunnel_id: logging.error("create_cloudflare_dns_record: Missing args."); return None
-    record_name = hostname; record_content = f"{tunnel_id}.cfargotunnel.com"; endpoint = f"/zones/{zone_id}/dns_records"; payload = {"type": "CNAME", "name": record_name, "content": record_content, "ttl": 1, "proxied": True }
+    """Creates a CNAME DNS record pointing to the tunnel if it doesn't exist."""
+    if not zone_id or not hostname or not tunnel_id:
+        logging.error("create_cloudflare_dns_record: Missing required arguments.")
+        return None
+
+    record_name = hostname
+    record_content = f"{tunnel_id}.cfargotunnel.com"
+    endpoint = f"/zones/{zone_id}/dns_records"
+    payload = {
+        "type": "CNAME",
+        "name": record_name,
+        "content": record_content,
+        "ttl": 1, # 1 = Automatic TTL
+        "proxied": True
+    }
+
     try:
+        # Check if it already exists with the correct content
         existing_id = find_dns_record_id(zone_id, hostname, tunnel_id)
-        if existing_id: logging.info(f"DNS CNAME for {hostname} in zone {zone_id} already exists. ID: {existing_id}."); return existing_id
+        if existing_id:
+            logging.info(f"DNS CNAME for {hostname} in zone {zone_id} already exists. ID: {existing_id}.")
+            return existing_id # Return existing ID
+
+        # If not found, create it
         logging.info(f"Creating DNS CNAME in zone {zone_id}: Name={record_name}, Content={record_content}, Proxied=True")
-        response_data = cf_api_request("POST", endpoint, json_data=payload); result = response_data.get("result", {}); new_record_id = result.get("id")
-        if new_record_id: logging.info(f"Successfully created DNS record for {hostname} in zone {zone_id}. New ID: {new_record_id}"); return new_record_id
-        else: logging.error(f"DNS record creation for {hostname} success but no ID returned: {result}"); return None
-    except requests.exceptions.RequestException as e: logging.error(f"API error creating DNS record for {hostname}: {e}"); return None
-    except Exception as e: logging.error(f"Unexpected error creating DNS record for {hostname}: {e}", exc_info=True); return None
+        response_data = cf_api_request("POST", endpoint, json_data=payload)
+        result = response_data.get("result", {})
+        new_record_id = result.get("id")
+        if new_record_id:
+            logging.info(f"Successfully created DNS record for {hostname} in zone {zone_id}. New ID: {new_record_id}")
+            return new_record_id
+        else:
+            # This shouldn't happen if success was true, but log defensively
+            logging.error(f"DNS record creation API call for {hostname} reported success but response missing ID: {result}")
+            return None
+    except requests.exceptions.RequestException as e:
+        logging.error(f"API error creating DNS record for {hostname}: {e}")
+        return None
+    except Exception as e:
+        logging.error(f"Unexpected error creating DNS record for {hostname}: {e}", exc_info=True)
+        return None
+
 
 def delete_cloudflare_dns_record(zone_id, hostname, tunnel_id):
-    if not zone_id or not hostname or not tunnel_id: logging.error("delete_cloudflare_dns_record: Missing args."); return False
+    """Deletes the specific CNAME DNS record pointing to the tunnel."""
+    if not zone_id or not hostname or not tunnel_id:
+        logging.error("delete_cloudflare_dns_record: Missing required arguments.")
+        return False
+
     dns_record_id = find_dns_record_id(zone_id, hostname, tunnel_id)
-    if not dns_record_id: logging.warning(f"DNS record for {hostname} in zone {zone_id} not found to delete."); return True
-    logging.info(f"Attempting to delete DNS record for {hostname} in zone {zone_id} (ID: {dns_record_id})"); endpoint = f"/zones/{zone_id}/dns_records/{dns_record_id}"
-    try: cf_api_request("DELETE", endpoint); logging.info(f"Successfully deleted DNS record for {hostname} (ID: {dns_record_id})."); return True
+    if not dns_record_id:
+        logging.warning(f"DNS record for {hostname} in zone {zone_id} (pointing to tunnel {tunnel_id}) not found to delete. Assuming success.")
+        return True # Treat as success if not found
+
+    logging.info(f"Attempting to delete DNS record for {hostname} in zone {zone_id} (ID: {dns_record_id})")
+    endpoint = f"/zones/{zone_id}/dns_records/{dns_record_id}"
+    try:
+        cf_api_request("DELETE", endpoint)
+        logging.info(f"Successfully deleted DNS record for {hostname} (ID: {dns_record_id}).")
+        return True
     except requests.exceptions.RequestException as e:
-        if e.response is not None and e.response.status_code == 404: logging.warning(f"DNS record {dns_record_id} for {hostname} not found during delete (404). Treating as success."); return True
-        logging.error(f"API error deleting DNS record {dns_record_id} for {hostname}: {e}"); return False
-    except Exception as e: logging.error(f"Unexpected error deleting DNS record {dns_record_id} for {hostname}: {e}", exc_info=True); return False
+        # Handle 404 specifically - it means the record was already gone
+        if e.response is not None and e.response.status_code == 404:
+             logging.warning(f"DNS record {dns_record_id} for {hostname} not found during delete attempt (404). Treating as success.")
+             return True
+        logging.error(f"API error deleting DNS record {dns_record_id} for {hostname}: {e}")
+        return False
+    except Exception as e:
+        logging.error(f"Unexpected error deleting DNS record {dns_record_id} for {hostname}: {e}", exc_info=True)
+        return False
+
 
 def update_cloudflare_config():
-    if not tunnel_state.get("id"): logging.warning("Cannot update CF config, tunnel ID missing."); return False
-    final_ingress_rules = None; needs_api_update = False
+    """Updates the Cloudflare tunnel ingress configuration if needed."""
+    if not tunnel_state.get("id"):
+        logging.warning("Cannot update CF config, tunnel ID missing.")
+        return False
+
+    final_ingress_rules = None
+    needs_api_update = False
+
     with state_lock:
-        logging.info("Checking for Cloudflare tunnel config updates..."); desired_ingress_rules = []; catch_all_rule = {"service": "http_status:404"}
+        logging.info("Checking for Cloudflare tunnel config updates...")
+        desired_ingress_rules = []
+        catch_all_rule = {"service": "http_status:404"} # Default catch-all
+
         for hostname, rule_details in managed_rules.items():
-            if rule_details.get("status") == "active": service = rule_details.get("service");
-            if service: desired_ingress_rules.append({"hostname": hostname, "service": service}); else: logging.warning(f"Rule {hostname} active but missing service. Skipping.")
-        desired_ingress_rules.sort(key=lambda x: x.get("hostname", "")); current_config = get_current_cf_config()
-        if current_config is None: logging.error("Failed fetch CF config, aborting update."); return False
+            if rule_details.get("status") == "active":
+                service = rule_details.get("service")
+                if service:
+                    desired_ingress_rules.append({"hostname": hostname, "service": service})
+                else:
+                    logging.warning(f"Rule {hostname} is active but missing 'service' detail. Skipping.")
+
+        # Sort desired rules for consistent comparison
+        desired_ingress_rules.sort(key=lambda x: x.get("hostname", ""))
+
+        logging.debug("Fetching current CF config for comparison...")
+        current_config = get_current_cf_config()
+        if current_config is None:
+            logging.error("Failed to fetch current CF config, aborting update check.")
+            return False
+
+        # Extract current rules, excluding the default catch-all
         current_cf_ingress = [r for r in current_config.get("ingress", []) if r.get("service") != catch_all_rule["service"]]
-        def rule_to_canonical(rule): items = sorted([(k, v) for k, v in rule.items() if k in ["hostname", "service"]]); return tuple(items)
-        try: current_cf_set = {rule_to_canonical(r) for r in current_cf_ingress if r.get("hostname") and r.get("service")}; desired_set = {rule_to_canonical(r) for r in desired_ingress_rules if r.get("hostname") and r.get("service")}
-        except Exception as e: logging.error(f"Error creating canonical rule sets: {e}", exc_info=True); return False
-        if current_cf_set == desired_set: logging.info("No changes detected in CF tunnel config. Skipping API update."); needs_api_update = False
-        else: logging.info("Change detected. Desired ingress rules differ from current CF config."); logging.debug(f"Current CF rules: {current_cf_set}"); logging.debug(f"Desired rules: {desired_set}"); needs_api_update = True; final_ingress_rules = desired_ingress_rules + [catch_all_rule]
+
+        # Compare rule sets canonically (convert dicts to sorted tuples of key-value pairs)
+        def rule_to_canonical(rule):
+            items = sorted([(k, v) for k, v in rule.items() if k in ["hostname", "service"]])
+            return tuple(items)
+
+        try:
+             current_cf_set = {rule_to_canonical(r) for r in current_cf_ingress if r.get("hostname") and r.get("service")}
+             desired_set = {rule_to_canonical(r) for r in desired_ingress_rules if r.get("hostname") and r.get("service")}
+        except Exception as e:
+             # Handle potential errors during set creation (e.g., unhashable types)
+             logging.error(f"Error creating canonical rule sets for comparison: {e}", exc_info=True)
+             return False
+
+        if current_cf_set == desired_set:
+            logging.info("No changes detected in CF tunnel config. Skipping API update.")
+            needs_api_update = False
+        else:
+            logging.info("Change detected. Desired ingress rules differ from current CF config.")
+            logging.debug(f"Current CF rules: {current_cf_set}")
+            logging.debug(f"Desired rules: {desired_set}")
+            needs_api_update = True
+            # Prepare the final list, including the catch-all rule
+            final_ingress_rules = desired_ingress_rules + [catch_all_rule]
+
+    # --- Perform API Update if needed ---
     if needs_api_update and final_ingress_rules is not None:
-        endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel/{tunnel_state['id']}/configurations"; payload = {"config": {"ingress": final_ingress_rules}}; last_exception = None
+        endpoint = f"/accounts/{CF_ACCOUNT_ID}/cfd_tunnel/{tunnel_state['id']}/configurations"
+        payload = {"config": {"ingress": final_ingress_rules}}
+        last_exception = None
+
+        # Retry logic
         for attempt in range(MAX_CF_UPDATE_RETRIES + 1):
             try:
                 logging.info(f"Attempting CF config push (Attempt {attempt + 1}/{MAX_CF_UPDATE_RETRIES + 1})...")
-                cf_api_request("PUT", endpoint, json_data=payload); logging.info("Successfully updated CF tunnel configuration via API.")
+                cf_api_request("PUT", endpoint, json_data=payload)
+                logging.info("Successfully updated CF tunnel configuration via API.")
                 cloudflared_agent_state["last_action_status"] = f"CF config updated successfully at {datetime.now(timezone.utc).isoformat()}"
-                if tunnel_state.get("error") and ("Failed update tunnel config" in tunnel_state["error"] or "API Error" in tunnel_state["error"]): tunnel_state["error"] = None
-                return True
+                # Clear previous API errors if update succeeds now
+                if tunnel_state.get("error") and ("Failed update tunnel config" in tunnel_state["error"] or "API Error" in tunnel_state["error"]):
+                     logging.info(f"Clearing previous API error related to config update: {tunnel_state['error']}")
+                     tunnel_state["error"] = None
+                return True # Success!
             except requests.exceptions.RequestException as e:
-                last_exception = e; status_code = e.response.status_code if e.response is not None else None; logging.warning(f"CF API update attempt {attempt + 1} failed: {e} (Status: {status_code})")
+                last_exception = e
+                status_code = e.response.status_code if e.response is not None else None
+                logging.warning(f"CF API update attempt {attempt + 1} failed: {e} (Status: {status_code})")
+
+                # Determine if retryable
                 is_retryable = isinstance(e, (requests.exceptions.ConnectionError, requests.exceptions.Timeout)) or status_code in [429, 500, 502, 503, 504]
+
                 if is_retryable and attempt < MAX_CF_UPDATE_RETRIES:
-                    wait_time = CF_UPDATE_RETRY_DELAY * (CF_UPDATE_BACKOFF_FACTOR ** attempt) * (1 + random.uniform(-0.2, 0.2)); wait_time = max(1, wait_time)
-                    if status_code == 429 and e.response is not None and e.response.headers.get("Retry-After"):
-                         try: wait_time = max(wait_time, int(e.response.headers["Retry-After"])); logging.info(f"Respecting Retry-After: {e.response.headers['Retry-After']}s")
-                         except ValueError: pass
+                    # Exponential backoff with jitter
+                    wait_time = CF_UPDATE_RETRY_DELAY * (CF_UPDATE_BACKOFF_FACTOR ** attempt)
+                    wait_time *= (1 + random.uniform(-0.2, 0.2)) # Jitter +/- 20%
+                    wait_time = max(1, wait_time) # Minimum 1 second wait
+
+                    # Respect Retry-After header if present (Code 429)
+                    if status_code == 429 and e.response is not None:
+                         retry_after = e.response.headers.get("Retry-After")
+                         if retry_after:
+                              try:
+                                   wait_time = max(wait_time, int(retry_after))
+                                   logging.info(f"Respecting Retry-After header: {retry_after} seconds.")
+                              except ValueError:
+                                   logging.warning(f"Could not parse Retry-After header value: {retry_after}")
+
                     logging.info(f"Retrying CF update in {wait_time:.1f} seconds...")
-                    if stop_event.wait(wait_time): logging.warning("Shutdown during CF update retry wait."); cloudflared_agent_state["last_action_status"] = "Error: CF update aborted (shutdown)."; tunnel_state["error"] = "Failed update: aborted retry"; return False; continue
-                else: logging.error(f"CF API update failed, won't retry (Retryable: {is_retryable}, Attempt: {attempt + 1})."); break
-            except Exception as e: last_exception = e; logging.error(f"Unexpected error during CF API update attempt {attempt + 1}: {e}", exc_info=True); break
-        logging.error(f"Failed to update CF tunnel config after {MAX_CF_UPDATE_RETRIES + 1} attempts."); error_message = f"Failed update tunnel config: {last_exception}"
-        cloudflared_agent_state["last_action_status"] = f"Error: {error_message}"; if not tunnel_state.get("error"): tunnel_state["error"] = error_message
+                    # Wait, but check if shutdown is requested
+                    if stop_event.wait(wait_time):
+                        logging.warning("Shutdown requested during CF update retry wait.")
+                        cloudflared_agent_state["last_action_status"] = "Error: CF update aborted (shutdown during retry)."
+                        tunnel_state["error"] = "Failed update tunnel config: aborted retry"
+                        return False
+                    continue # Go to next attempt
+                else:
+                    # Not retryable or max retries reached
+                    logging.error(f"CF API update failed permanently (Retryable: {is_retryable}, Attempt: {attempt + 1}).")
+                    break # Exit retry loop
+            except Exception as e:
+                # Catch unexpected errors during the attempt
+                last_exception = e
+                logging.error(f"Unexpected error during CF API update attempt {attempt + 1}: {e}", exc_info=True)
+                break # Exit retry loop
+
+        # If loop finished without success
+        logging.error(f"Failed to update CF tunnel config after {MAX_CF_UPDATE_RETRIES + 1} attempts.")
+        error_message = f"Failed update tunnel config: {last_exception}"
+        cloudflared_agent_state["last_action_status"] = f"Error: {error_message}"
+        if not tunnel_state.get("error"): # Don't overwrite more specific errors
+            tunnel_state["error"] = error_message
         return False
-    elif needs_api_update and final_ingress_rules is None: logging.error("Internal error: needs update but final rules not set."); return False
-    else: return True
+
+    elif needs_api_update and final_ingress_rules is None:
+        # Should not happen if logic is correct
+        logging.error("Internal error: update needed but final ingress rules were not generated.")
+        return False
+    else:
+        # No update was needed
+        return True
+
 
 def process_container_start(container):
-    if not container: return; container_id = None; container_name = "Unknown"
+    """Processes a container start event based on labels."""
+    if not container: return
+    container_id = None
+    container_name = "Unknown"
     try:
-        container_id = container.id;
-        try: container.reload(); container_name = container.name
-        except NotFound: logging.warning(f"Container {container_id[:12]} not found processing start (stopped quickly?)."); return
-        except APIError as e: logging.error(f"Docker API error reloading container {container_id[:12]}: {e}"); return
+        container_id = container.id
+        # Reload container state to ensure labels are fresh
+        try:
+             container.reload()
+             container_name = container.name # Get name after reload
+        except NotFound:
+             logging.warning(f"Container {container_id[:12]} not found processing start (likely stopped very quickly?).")
+             return
+        except APIError as e:
+             logging.error(f"Docker API error reloading container {container_id[:12]}: {e}")
+             return # Cannot proceed without reliable info
 
-        labels = container.labels; enabled_label = f"{LABEL_PREFIX}.enable"; hostname_label = f"{LABEL_PREFIX}.hostname"; service_label = f"{LABEL_PREFIX}.service"; zone_name_label = f"{LABEL_PREFIX}.zonename"
-        is_enabled = labels.get(enabled_label, "false").lower() in ["true", "1", "t", "yes"]; hostname = labels.get(hostname_label); service = labels.get(service_label); zone_name = labels.get(zone_name_label)
-        if not is_enabled: logging.debug(f"Ignoring start: {container_name}: '{enabled_label}' not true."); return
-        if not hostname or not service: logging.warning(f"Ignoring start: {container_name}: Missing '{hostname_label}' or '{service_label}'."); return
-        if not re.match(r"^[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)+$", hostname): logging.warning(f"Ignoring start: {container_name}: Invalid hostname format '{hostname}'."); return
-        if not (re.match(r"^(https?|tcp|unix)://", service) or re.match(r"^[a-zA-Z0-9._-]+:\d+$", service)): logging.warning(f"Ignoring start: {container_name}: Invalid service format '{service}'."); return
+        labels = container.labels
+        enabled_label = f"{LABEL_PREFIX}.enable"
+        hostname_label = f"{LABEL_PREFIX}.hostname"
+        service_label = f"{LABEL_PREFIX}.service"
+        zone_name_label = f"{LABEL_PREFIX}.zonename" # Label for specific zone
+
+        is_enabled = labels.get(enabled_label, "false").lower() in ["true", "1", "t", "yes"]
+        hostname = labels.get(hostname_label)
+        service = labels.get(service_label)
+        zone_name = labels.get(zone_name_label) # Read optional zone name label
+
+        if not is_enabled:
+            logging.debug(f"Ignoring start: {container_name} ({container_id[:12]}): '{enabled_label}' not true.")
+            return
+        if not hostname or not service:
+            logging.warning(f"Ignoring start: {container_name} ({container_id[:12]}): Missing '{hostname_label}' or '{service_label}'.")
+            return
+        # Basic Hostname validation
+        if not re.match(r"^[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)+$", hostname):
+             logging.warning(f"Ignoring start: {container_name} ({container_id[:12]}): Invalid hostname format '{hostname}'.")
+             return
+        # Basic Service validation (URL-like or host:port)
+        if not (re.match(r"^(https?|tcp|unix)://", service) or re.match(r"^[a-zA-Z0-9._-]+:\d+$", service)):
+             logging.warning(f"Ignoring start: {container_name} ({container_id[:12]}): Invalid service format '{service}'. Needs protocol (http/https/tcp/unix) or host:port.")
+             return
+
+        # --- Determine Zone ID ---
         target_zone_id = None
-        if zone_name: logging.info(f"Container {container_name} specified zone name: '{zone_name}'. Looking up ID."); target_zone_id = get_zone_id_from_name(zone_name);
-        if not target_zone_id: logging.error(f"Failed to find Zone ID for specified name '{zone_name}' for container {container_name}.")
-        else: logging.debug(f"Container {container_name} did not specify zone name. Using default Zone ID."); target_zone_id = CF_ZONE_ID
-        if not target_zone_id: logging.error(f"Cannot manage DNS for {hostname} (container {container_name}): No valid Zone ID found."); return
-        logging.info(f"Managing {hostname} (from {container_name}) in Zone ID: {target_zone_id}"); needs_cf_update = False; state_changed_locally = False
+        if zone_name:
+            logging.info(f"Container {container_name} specified zone name: '{zone_name}'. Looking up ID.")
+            target_zone_id = get_zone_id_from_name(zone_name)
+            if not target_zone_id:
+                logging.error(f"Failed to find Zone ID for specified name '{zone_name}' for container {container_name}. Cannot manage DNS for {hostname}.")
+                # Do not proceed without a valid zone ID if specified
+                return
+        else:
+            logging.debug(f"Container {container_name} did not specify zone name. Using default Zone ID if available.")
+            target_zone_id = CF_ZONE_ID # Use default from environment (might be None)
+
+        if not target_zone_id:
+             logging.error(f"Cannot manage DNS for {hostname} (container {container_name}): No valid Zone ID found (label lookup failed and no default CF_ZONE_ID set?).")
+             # Stop processing this container if no zone ID could be determined
+             return
+
+        logging.info(f"Managing {hostname} (from {container_name}) in Zone ID: {target_zone_id}")
+
+        # --- Update State ---
+        needs_cf_update = False
+        state_changed_locally = False
         with state_lock:
             existing_rule = managed_rules.get(hostname)
             if existing_rule:
+                # Rule already exists, check for changes or reactivation
                 zone_id_changed = existing_rule.get("zone_id") != target_zone_id
-                if existing_rule.get("status") == "pending_deletion": logging.info(f"Rule for {hostname} was pending deletion. Reactivating."); existing_rule["status"] = "active"; existing_rule["delete_at"] = None; existing_rule["service"] = service; existing_rule["container_id"] = container_id; existing_rule["zone_id"] = target_zone_id; state_changed_locally = True; needs_cf_update = True; if zone_id_changed: logging.info(f"Zone ID for reactivated rule {hostname} updated to {target_zone_id}.")
+
+                if existing_rule.get("status") == "pending_deletion":
+                    logging.info(f"Rule for {hostname} was pending deletion. Reactivating.")
+                    existing_rule["status"] = "active"
+                    existing_rule["delete_at"] = None
+                    existing_rule["service"] = service # Update service/container in case they changed
+                    existing_rule["container_id"] = container_id
+                    existing_rule["zone_id"] = target_zone_id # Update stored zone ID
+                    state_changed_locally = True
+                    needs_cf_update = True # Need to update CF ingress config
+                    if zone_id_changed:
+                        logging.info(f"Zone ID for reactivated rule {hostname} updated to {target_zone_id}.")
                 elif existing_rule.get("status") == "active":
-                    service_changed = existing_rule.get("service") != service; container_changed = existing_rule.get("container_id") != container_id
-                    if container_changed: logging.info(f"Updating container ID for active rule {hostname}: '{existing_rule.get('container_id')[:12]}' -> '{container_id[:12]}'."); existing_rule["container_id"] = container_id; state_changed_locally = True
-                    if service_changed: logging.info(f"Updating service for active rule {hostname}: '{existing_rule.get('service')}' -> '{service}'."); existing_rule["service"] = service; state_changed_locally = True; needs_cf_update = True
-                    if zone_id_changed: logging.warning(f"Zone ID for active rule {hostname} changed ('{existing_rule.get('zone_id')}' -> '{target_zone_id}'). DNS record might be stale in old zone."); existing_rule["zone_id"] = target_zone_id; state_changed_locally = True; needs_cf_update = True
-            else: logging.info(f"Adding new active rule for hostname: {hostname}"); managed_rules[hostname] = {"service": service, "container_id": container_id, "status": "active", "delete_at": None, "zone_id": target_zone_id}; state_changed_locally = True; needs_cf_update = True
+                    # Rule is already active, check if details changed
+                    service_changed = existing_rule.get("service") != service
+                    container_changed = existing_rule.get("container_id") != container_id
+
+                    if container_changed:
+                        logging.info(f"Updating container ID for active rule {hostname}: '{existing_rule.get('container_id')[:12]}' -> '{container_id[:12]}'.")
+                        existing_rule["container_id"] = container_id
+                        state_changed_locally = True
+                        # Container ID change doesn't require CF update unless service also changed
+                    if service_changed:
+                         logging.info(f"Updating service for active rule {hostname}: '{existing_rule.get('service')}' -> '{service}'.")
+                         existing_rule["service"] = service
+                         state_changed_locally = True
+                         needs_cf_update = True # Service change requires CF update
+                    if zone_id_changed:
+                         # This is tricky. DNS might be stale in the old zone. We update our state.
+                         # DNS creation/update in the new zone happens after the potential CF update.
+                         logging.warning(f"Zone ID for active rule {hostname} changed ('{existing_rule.get('zone_id')}' -> '{target_zone_id}'). DNS in old zone may be stale if cleanup failed.")
+                         existing_rule["zone_id"] = target_zone_id
+                         state_changed_locally = True
+                         # We still need to ensure DNS exists in the *new* zone, trigger update path
+                         needs_cf_update = True
+            else:
+                # New rule for a new hostname
+                logging.info(f"Adding new active rule for hostname: {hostname}")
+                managed_rules[hostname] = {
+                    "service": service,
+                    "container_id": container_id,
+                    "status": "active",
+                    "delete_at": None,
+                    "zone_id": target_zone_id # Store the determined Zone ID
+                }
+                state_changed_locally = True
+                needs_cf_update = True
+
             if state_changed_locally:
                 logging.debug(f"Saving state after processing start for {hostname}.")
                 save_state()
-                send_sse_event('rules_update') # Signal frontend after state saved
+
+        # --- Update Cloudflare (Tunnel Config and DNS) ---
         if needs_cf_update:
             logging.info(f"Triggering Cloudflare config update due to change for {hostname}.")
-            if update_cloudflare_config(): logging.info(f"Tunnel config update successful for {hostname}.")
-            if tunnel_state.get("id"): dns_record_id = create_cloudflare_dns_record(target_zone_id, hostname, tunnel_state["id"]);
-            if dns_record_id: logging.info(f"DNS record management in zone {target_zone_id} successful for {hostname}.")
-            else: logging.error(f"CRITICAL: Tunnel config updated for {hostname} but failed to create/verify DNS record in zone {target_zone_id}!"); cloudflared_agent_state["last_action_status"] = f"Error: Failed creating DNS for {hostname} in zone {target_zone_id}."
-            else: logging.error("Missing Tunnel ID - cannot manage DNS record.")
-            else: logging.error(f"Failed to update Cloudflare tunnel config after processing start for {hostname}. DNS record not managed.")
-        elif state_changed_locally: logging.debug(f"Local state updated for {hostname} (e.g., container ID), no Cloudflare config change needed.")
-    except NotFound: logging.warning(f"Container {container_name} ({container_id[:12] if container_id else 'Unknown'}) not found during start processing.")
-    except APIError as e: logging.error(f"Docker API error processing container start ({container_name}): {e}", exc_info=True)
-    except Exception as e: logging.error(f"Unexpected error processing container start ({container_name}): {e}", exc_info=True)
+            # Attempt to update the tunnel ingress configuration first
+            if update_cloudflare_config():
+                logging.info(f"Tunnel config update successful for {hostname}.")
+                # If tunnel config update succeeded, manage the DNS record
+                if tunnel_state.get("id"): # Ensure tunnel ID exists
+                    dns_record_id = create_cloudflare_dns_record(target_zone_id, hostname, tunnel_state["id"]) # Use target_zone_id
+                    if dns_record_id:
+                         logging.info(f"DNS record management in zone {target_zone_id} successful for {hostname}.")
+                    else:
+                         # Log critical error if DNS fails after tunnel config update
+                         logging.error(f"CRITICAL: Tunnel config updated for {hostname} but failed to create/verify DNS record in zone {target_zone_id}!")
+                         cloudflared_agent_state["last_action_status"] = f"Error: Failed creating DNS for {hostname} in zone {target_zone_id}."
+                else:
+                     logging.error("Missing Tunnel ID - cannot manage DNS record for {hostname}.")
+            else:
+                # Log error if tunnel config update fails
+                logging.error(f"Failed to update Cloudflare tunnel config after processing start for {hostname}. DNS record not managed.")
+        elif state_changed_locally:
+             # Only container ID might have changed, no CF update needed
+             logging.debug(f"Local state updated for {hostname} (e.g., container ID), no Cloudflare config change needed.")
+
+    except NotFound:
+        # Handles case where container disappears between event and get/reload
+        logging.warning(f"Container {container_name} ({container_id[:12] if container_id else 'Unknown'}) not found during start processing.")
+    except APIError as e:
+        logging.error(f"Docker API error processing container start ({container_name}): {e}", exc_info=True)
+    except Exception as e:
+        logging.error(f"Unexpected error processing container start ({container_name}): {e}", exc_info=True)
+
 
 def schedule_container_stop(container_id):
-    if not container_id: return; logging.info(f"Processing stop event for container {container_id[:12]}."); hostname_to_schedule = None; state_changed = False
+    """Schedules a rule for deletion when its container stops."""
+    if not container_id: return
+    logging.info(f"Processing stop event for container {container_id[:12]}.")
+    hostname_to_schedule = None
+    state_changed = False
     with state_lock:
+        # Find the rule managed by this container ID
         for hn, details in managed_rules.items():
-            if details.get("container_id") == container_id and details.get("status") == "active": hostname_to_schedule = hn; break
+            # Only schedule if the rule is currently active
+            if details.get("container_id") == container_id and details.get("status") == "active":
+                hostname_to_schedule = hn
+                break
+
         if hostname_to_schedule:
             logging.info(f"Container {container_id[:12]} managed hostname '{hostname_to_schedule}'. Marking rule for deletion.")
             rule = managed_rules[hostname_to_schedule]
-            if rule.get("status") != "pending_deletion": rule["status"] = "pending_deletion"; rule["delete_at"] = datetime.now(timezone.utc) + timedelta(seconds=GRACE_PERIOD_SECONDS); logging.info(f"Rule for {hostname_to_schedule} scheduled for deletion at {rule['delete_at'].isoformat()}"); state_changed = True
-            else: logging.info(f"Rule for {hostname_to_schedule} was already pending deletion.")
-        else: logging.info(f"Stop event for {container_id[:12]}, but it didn't manage an active rule in current state.")
+            if rule.get("status") != "pending_deletion":
+                 rule["status"] = "pending_deletion"
+                 rule["delete_at"] = datetime.now(timezone.utc) + timedelta(seconds=GRACE_PERIOD_SECONDS)
+                 logging.info(f"Rule for {hostname_to_schedule} scheduled for deletion at {rule['delete_at'].isoformat()}")
+                 state_changed = True
+            else:
+                 # Already pending, don't reset timer
+                 logging.info(f"Rule for {hostname_to_schedule} was already pending deletion.")
+        else:
+            logging.info(f"Stop event for {container_id[:12]}, but it didn't manage an active rule in current state.")
+
         if state_changed:
             save_state()
-            send_sse_event('rules_update') # Signal potential change (status goes to pending)
+
 
 def docker_event_listener():
-    if not docker_client: logging.error("Docker client unavailable, event listener cannot start."); return
-    logging.info("Starting Docker event listener..."); error_count = 0; max_errors = 5
+    """Listens for Docker start/stop events and processes them."""
+    if not docker_client:
+        logging.error("Docker client unavailable, event listener cannot start.")
+        return
+
+    logging.info("Starting Docker event listener...")
+    error_count = 0
+    max_errors = 5 # Limit consecutive errors before stopping listener
+
     while not stop_event.is_set() and error_count < max_errors:
         try:
-            logging.info("Connecting to Docker event stream..."); events = docker_client.events(decode=True, since=int(time.time()))
-            logging.info("Successfully connected to Docker event stream."); error_count = 0
+            logging.info("Connecting to Docker event stream...")
+            # Get events since listener started to avoid processing old events on reconnect
+            events = docker_client.events(decode=True, since=int(time.time()))
+            logging.info("Successfully connected to Docker event stream.")
+            error_count = 0 # Reset error count on successful connection
+
             for event in events:
-                if stop_event.is_set(): logging.info("Stop event received, exiting listener loop."); break
-                ev_type = event.get("Type"); action = event.get("Action"); actor = event.get("Actor", {}); cont_id = actor.get("ID")
+                if stop_event.is_set():
+                    logging.info("Stop event received, exiting listener loop.")
+                    break
+
+                ev_type = event.get("Type")
+                action = event.get("Action")
+                actor = event.get("Actor", {})
+                cont_id = actor.get("ID")
+
                 logging.debug(f"Docker Event: Type={ev_type}, Action={action}, ID={cont_id[:12] if cont_id else 'N/A'}")
+
                 if ev_type == "container" and cont_id:
                     if action == "start":
-                        try: time.sleep(0.5); container = docker_client.containers.get(cont_id); process_container_start(container)
-                        except NotFound: logging.warning(f"Container {cont_id[:12]} not found shortly after 'start' event.")
-                        except APIError as e: logging.error(f"Docker API error processing start event for {cont_id[:12]}: {e}")
-                        except Exception as e: logging.error(f"Unexpected error processing start event for {cont_id[:12]}: {e}", exc_info=True)
+                        try:
+                            # Use a short delay to allow container info (like labels) to stabilize
+                            time.sleep(0.5)
+                            container = docker_client.containers.get(cont_id)
+                            process_container_start(container)
+                        except NotFound:
+                            # Container might be very short-lived
+                            logging.warning(f"Container {cont_id[:12]} not found shortly after 'start' event.")
+                        except APIError as e:
+                            logging.error(f"Docker API error processing start event for {cont_id[:12]}: {e}")
+                        except Exception as e:
+                            logging.error(f"Unexpected error processing start event for {cont_id[:12]}: {e}", exc_info=True)
                     elif action in ["stop", "die", "destroy", "kill"]:
-                         try: schedule_container_stop(cont_id)
-                         except Exception as e: logging.error(f"Unexpected error processing stop/die/destroy event for {cont_id[:12]}: {e}", exc_info=True)
-        except requests.exceptions.ConnectionError as e: error_count += 1; logging.error(f"Docker listener connection error: {e}. Reconnecting ({error_count}/{max_errors})..."); stop_event.wait(min(30, 5 * error_count))
-        except APIError as e: error_count += 1; logging.error(f"Docker listener API error: {e}. Reconnecting ({error_count}/{max_errors})..."); stop_event.wait(min(30, 5 * error_count))
-        except Exception as e: error_count += 1; logging.error(f"Unexpected error in Docker event listener: {e}. Reconnecting ({error_count}/{max_errors})...", exc_info=True); stop_event.wait(min(30, 5 * error_count))
-        if stop_event.is_set(): break
-    if error_count >= max_errors: logging.error("Docker event listener stopping after multiple consecutive errors.")
+                        # Note: destroy might not always have the ID easily available if already gone
+                         try:
+                             schedule_container_stop(cont_id)
+                         except Exception as e:
+                             logging.error(f"Unexpected error processing stop/die/destroy event for {cont_id[:12]}: {e}", exc_info=True)
+
+        except requests.exceptions.ConnectionError as e:
+            error_count += 1
+            logging.error(f"Docker listener connection error: {e}. Reconnecting ({error_count}/{max_errors})...")
+            stop_event.wait(min(30, 5 * error_count)) # Exponential backoff for reconnect
+        except APIError as e:
+             error_count += 1
+             logging.error(f"Docker listener API error: {e}. Reconnecting ({error_count}/{max_errors})...")
+             stop_event.wait(min(30, 5 * error_count))
+        except Exception as e:
+            # Catch other unexpected errors in the listener loop
+            error_count += 1
+            logging.error(f"Unexpected error in Docker event listener: {e}. Reconnecting ({error_count}/{max_errors})...", exc_info=True)
+            stop_event.wait(min(30, 5 * error_count))
+
+        # Ensure loop exits if stop event is set during wait
+        if stop_event.is_set():
+            break
+
+    if error_count >= max_errors:
+        logging.error("Docker event listener stopping after multiple consecutive errors.")
     logging.info("Docker event listener stopped.")
 
+
 def cleanup_expired_rules():
-    logging.info("Starting cleanup task...");
+    """Periodically checks for and cleans up expired rules."""
+    logging.info("Starting cleanup task...")
     while not stop_event.is_set():
         next_check_time = time.time() + CLEANUP_INTERVAL_SECONDS
         try:
-            logging.debug("Running cleanup check for expired rules..."); rules_to_delete = {}; now_utc = datetime.now(timezone.utc); state_changed_in_cleanup = False
+            logging.debug("Running cleanup check for expired rules...")
+            rules_to_delete = {} # hostname -> zone_id
+            now_utc = datetime.now(timezone.utc)
+            state_changed_in_cleanup = False
+
             with state_lock:
                 for hostname, details in managed_rules.items():
                     if details.get("status") == "pending_deletion":
-                        delete_at = details.get("delete_at"); is_expired = False
+                        delete_at = details.get("delete_at")
+                        is_expired = False
                         if isinstance(delete_at, datetime):
-                             if delete_at.astimezone(timezone.utc) <= now_utc: is_expired = True
-                        else: logging.warning(f"Rule {hostname} pending delete but delete_at invalid: {delete_at}. Deleting immediately."); is_expired = True
-                        if is_expired: zone_id_for_delete = details.get("zone_id", CF_ZONE_ID);
-                        if not zone_id_for_delete: logging.error(f"Cannot schedule DNS deletion for expired rule {hostname}: Zone ID missing and no default set.")
-                        else: rules_to_delete[hostname] = zone_id_for_delete; logging.info(f"Rule for {hostname} in zone {zone_id_for_delete} expired. Scheduling for full deletion.")
+                             # Ensure comparison is timezone-aware
+                             delete_at_utc = delete_at.astimezone(timezone.utc)
+                             if delete_at_utc <= now_utc:
+                                 is_expired = True
+                        else:
+                             # Invalid delete_at timestamp, treat as immediately expired
+                             logging.warning(f"Rule {hostname} pending delete but has invalid delete_at timestamp: {delete_at}. Deleting immediately.")
+                             is_expired = True
+
+                        if is_expired:
+                            # Get zone ID from rule, fallback to default if missing
+                            zone_id_for_delete = details.get("zone_id", CF_ZONE_ID)
+                            if not zone_id_for_delete:
+                                logging.error(f"Cannot schedule DNS deletion for expired rule {hostname}: Zone ID is missing in state and no default CF_ZONE_ID is set.")
+                            else:
+                                rules_to_delete[hostname] = zone_id_for_delete
+                                logging.info(f"Rule for {hostname} in zone {zone_id_for_delete} expired. Scheduling for full deletion.")
+
+            # --- Perform Deletions Outside Lock ---
             if rules_to_delete:
-                logging.info(f"Processing cleanup for hostnames: {list(rules_to_delete.keys())}"); processed_hostnames_for_cf_update = []; dns_delete_success_all = True
+                logging.info(f"Processing cleanup for hostnames: {list(rules_to_delete.keys())}")
+                processed_hostnames_for_cf_update = []
+                dns_delete_success_all = True
+
+                # Step 1: Delete DNS records first
                 for hostname, zone_id in rules_to_delete.items():
-                    if tunnel_state.get("id"): logging.info(f"Attempting DNS record deletion for expired rule: {hostname} in zone {zone_id}");
-                    if delete_cloudflare_dns_record(zone_id, hostname, tunnel_state["id"]): processed_hostnames_for_cf_update.append(hostname)
-                    else: logging.error(f"Failed to delete DNS record for {hostname} in zone {zone_id}."); dns_delete_success_all = False; processed_hostnames_for_cf_update.append(hostname)
-                    else: logging.error(f"Cannot delete DNS for expired rule {hostname}: Missing Tunnel ID."); dns_delete_success_all = False; processed_hostnames_for_cf_update.append(hostname)
+                    if tunnel_state.get("id"):
+                         logging.info(f"Attempting DNS record deletion for expired rule: {hostname} in zone {zone_id}")
+                         if delete_cloudflare_dns_record(zone_id, hostname, tunnel_state["id"]):
+                              # Only add to list for tunnel config update if DNS delete was attempted/successful
+                              processed_hostnames_for_cf_update.append(hostname)
+                         else:
+                              logging.error(f"Failed to delete DNS record for {hostname} in zone {zone_id}. Tunnel config update will proceed, but DNS record may remain stale.")
+                              dns_delete_success_all = False
+                              processed_hostnames_for_cf_update.append(hostname) # Still try to remove from tunnel config
+                    else:
+                         logging.error(f"Cannot delete DNS for expired rule {hostname}: Missing Tunnel ID.")
+                         dns_delete_success_all = False
+                         # Cannot proceed with DNS delete, skip adding to tunnel update list for now?
+                         # Or add it anyway assuming tunnel config should be source of truth? Let's add it.
+                         processed_hostnames_for_cf_update.append(hostname)
+
+
+                # Step 2: Update Cloudflare tunnel config (removes based on active rules in state)
+                # The update_cloudflare_config function inherently only includes "active" rules.
+                # So, triggering it effectively removes the now-expired (pending_deletion) rules.
                 if processed_hostnames_for_cf_update:
                     logging.info(f"Attempting Cloudflare tunnel config update after processing DNS deletions for: {processed_hostnames_for_cf_update}")
-                    if update_cloudflare_config(): logging.info(f"Cloudflare tunnel config updated. Removing rules from local state: {processed_hostnames_for_cf_update}")
-                    with state_lock:
-                        deleted_count = 0
-                        for hostname in processed_hostnames_for_cf_update:
-                            if hostname in managed_rules and managed_rules[hostname].get("status") == "pending_deletion": del managed_rules[hostname]; deleted_count += 1; state_changed_in_cleanup = True
-                            else: logging.warning(f"Rule {hostname} was scheduled for removal but not found or not pending when removing from state.")
-                        logging.info(f"Removed {deleted_count} rules from local state.");
-                        if state_changed_in_cleanup:
-                            save_state()
-                            send_sse_event('rules_update') # Signal after successful cleanup and save
-                    else: logging.error("Failed to update Cloudflare tunnel config during rule cleanup. Rules remain in local state. Will retry on next cycle.")
-                else: logging.info("No hostnames eligible for tunnel config update after DNS processing during cleanup.")
-            else: logging.debug("No expired rules found requiring cleanup.")
-        except Exception as e: logging.error(f"Error in cleanup task loop: {e}", exc_info=True)
-        wait_duration = max(0, next_check_time - time.time()); stop_event.wait(wait_duration)
+                    if update_cloudflare_config():
+                        logging.info(f"Cloudflare tunnel config updated. Removing rules from local state: {processed_hostnames_for_cf_update}")
+                        # Step 3: Remove from local state ONLY if CF update succeeded
+                        with state_lock:
+                            deleted_count = 0
+                            for hostname in processed_hostnames_for_cf_update:
+                                # Double-check it exists and is still pending before deleting
+                                if hostname in managed_rules and managed_rules[hostname].get("status") == "pending_deletion":
+                                    del managed_rules[hostname]
+                                    deleted_count += 1
+                                    state_changed_in_cleanup = True
+                                else:
+                                    # This might happen if another process modified state concurrently, though unlikely with lock
+                                    logging.warning(f"Rule {hostname} was scheduled for removal but not found or not pending when removing from state.")
+                            logging.info(f"Removed {deleted_count} rules from local state.")
+                            if state_changed_in_cleanup:
+                                save_state()
+                    else:
+                        # If CF update failed, leave rules in local state (as pending_deletion)
+                        # They will be picked up by the next cleanup cycle.
+                        logging.error("Failed to update Cloudflare tunnel config during rule cleanup. Rules remain in local state. Will retry on next cycle.")
+                else:
+                     logging.info("No hostnames eligible for tunnel config update after DNS processing during cleanup.")
+
+            else:
+                logging.debug("No expired rules found requiring cleanup.")
+
+        except Exception as e:
+            logging.error(f"Error in cleanup task loop: {e}", exc_info=True)
+
+        # Wait until the next check time, respecting the stop event
+        wait_duration = max(0, next_check_time - time.time())
+        stop_event.wait(wait_duration)
+
     logging.info("Cleanup task stopped.")
 
+
 def reconcile_state():
-    if not docker_client: logging.warning("Docker client unavailable, skipping reconciliation."); return
-    if not tunnel_state.get("id"): logging.warning("Tunnel not initialized, skipping reconciliation."); return
-    logging.info("Starting state reconciliation..."); needs_cf_update = False; state_changed_locally = False
+    """Compares Docker state, local state, and Cloudflare state, making necessary adjustments."""
+    if not docker_client:
+        logging.warning("Docker client unavailable, skipping reconciliation.")
+        return
+    if not tunnel_state.get("id"):
+        logging.warning("Tunnel not initialized, skipping reconciliation.")
+        return
+
+    logging.info("Starting state reconciliation...")
+    needs_cf_update = False
+    state_changed_locally = False
     try:
-        running_labeled_containers = {}
+        # --- Get Current Docker State ---
+        running_labeled_containers = {} # hostname -> {details}
         try:
-             containers = docker_client.containers.list(sparse=False)
+             containers = docker_client.containers.list(sparse=False) # Get details needed for labels
              logging.debug(f"[Reconcile] Found {len(containers)} running containers.")
              for c in containers:
                  try:
-                     labels = c.labels; container_id = c.id; container_name = c.name; enabled = labels.get(f"{LABEL_PREFIX}.enable", "false").lower() in ["true", "1", "t", "yes"]; hostname = labels.get(f"{LABEL_PREFIX}.hostname"); service = labels.get(f"{LABEL_PREFIX}.service"); zone_name = labels.get(f"{LABEL_PREFIX}.zonename")
+                     labels = c.labels
+                     container_id = c.id
+                     container_name = c.name
+                     enabled = labels.get(f"{LABEL_PREFIX}.enable", "false").lower() in ["true", "1", "t", "yes"]
+                     hostname = labels.get(f"{LABEL_PREFIX}.hostname")
+                     service = labels.get(f"{LABEL_PREFIX}.service")
+                     zone_name = labels.get(f"{LABEL_PREFIX}.zonename") # Read zone name label
+
                      if enabled and hostname and service:
+                         # Basic validation (same as in process_container_start)
                          if not re.match(r"^[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)+$", hostname): continue
                          if not (re.match(r"^(https?|tcp|unix)://", service) or re.match(r"^[a-zA-Z0-9._-]+:\d+$", service)): continue
-                         if hostname in running_labeled_containers: logging.warning(f"[Reconcile] Duplicate hostname label '{hostname}' found on running containers: {container_name} and {running_labeled_containers[hostname]['container_name']}. Using latest found: {container_name}.")
-                         running_labeled_containers[hostname] = {"service": service, "container_id": container_id, "container_name": container_name, "zone_name": zone_name}
-                 except (NotFound, APIError) as e: logging.warning(f"[Reconcile] Docker error processing container {c.id[:12]}: {e}. Skipping this container."); continue
+
+                         if hostname in running_labeled_containers:
+                              # Handle potential duplicate hostnames (e.g., during restarts)
+                              logging.warning(f"[Reconcile] Duplicate hostname label '{hostname}' found on running containers: {container_name} and {running_labeled_containers[hostname]['container_name']}. Using latest found: {container_name}.")
+                         running_labeled_containers[hostname] = {
+                             "service": service,
+                             "container_id": container_id,
+                             "container_name": container_name,
+                             "zone_name": zone_name # Store zone name from label
+                         }
+                 except (NotFound, APIError) as e:
+                      # Container might disappear during list iteration
+                      logging.warning(f"[Reconcile] Docker error processing container {c.id[:12]}: {e}. Skipping this container.");
+                      continue
              logging.info(f"[Reconcile] Found {len(running_labeled_containers)} running containers with valid Dockflare labels.")
-        except (APIError, requests.exceptions.ConnectionError) as e: logging.error(f"[Reconcile] Docker error listing containers: {e}. Aborting reconciliation."); return
+        except (APIError, requests.exceptions.ConnectionError) as e:
+             logging.error(f"[Reconcile] Docker error listing containers: {e}. Aborting reconciliation.");
+             return
+
+        # --- Compare Docker State with Local State (under lock) ---
         with state_lock:
-            logging.debug("[Reconcile] Acquired state lock for comparison."); now_utc = datetime.now(timezone.utc); managed_hostnames = set(managed_rules.keys()); running_hostnames = set(running_labeled_containers.keys()); hostnames_requiring_dns_check = []
+            logging.debug("[Reconcile] Acquired state lock for comparison.")
+            now_utc = datetime.now(timezone.utc)
+            managed_hostnames = set(managed_rules.keys())
+            running_hostnames = set(running_labeled_containers.keys())
+            hostnames_requiring_dns_check = [] # Track hostnames needing DNS verification/creation
+
+            # 1. Check running containers against managed rules
             for hostname, running_details in running_labeled_containers.items():
+                # Determine target zone ID for this running container
                 target_zone_id = get_zone_id_from_name(running_details.get("zone_name")) if running_details.get("zone_name") else CF_ZONE_ID
-                if not target_zone_id: logging.error(f"[Reconcile] Skipping management for running container {running_details['container_name']} ({hostname}): No valid Zone ID determined."); continue
+                if not target_zone_id:
+                     logging.error(f"[Reconcile] Skipping management for running container {running_details['container_name']} ({hostname}): No valid Zone ID determined.")
+                     continue # Skip to next container if zone cannot be determined
+
                 if hostname in managed_rules:
-                    rule = managed_rules[hostname]; zone_id_changed = rule.get("zone_id") != target_zone_id
-                    if rule.get("status") == "pending_deletion": logging.info(f"[Reconcile] Hostname {hostname} is running again, reactivating pending rule."); rule["status"] = "active"; rule["delete_at"] = None; rule["service"] = running_details["service"]; rule["container_id"] = running_details["container_id"]; rule["zone_id"] = target_zone_id; state_changed_locally = True; needs_cf_update = True; hostnames_requiring_dns_check.append(hostname); if zone_id_changed: logging.info(f"[Reconcile] Zone ID for reactivated rule {hostname} updated to {target_zone_id}.")
+                    # Rule exists, check status and details
+                    rule = managed_rules[hostname]
+                    zone_id_changed = rule.get("zone_id") != target_zone_id
+
+                    if rule.get("status") == "pending_deletion":
+                        # Container is running again for a rule pending deletion - reactivate
+                        logging.info(f"[Reconcile] Hostname {hostname} is running again, reactivating pending rule.")
+                        rule["status"] = "active"; rule["delete_at"] = None
+                        rule["service"] = running_details["service"]; rule["container_id"] = running_details["container_id"]
+                        rule["zone_id"] = target_zone_id # Update zone ID
+                        state_changed_locally = True; needs_cf_update = True
+                        hostnames_requiring_dns_check.append(hostname) # Ensure DNS exists
+                        if zone_id_changed: logging.info(f"[Reconcile] Zone ID for reactivated rule {hostname} updated to {target_zone_id}.")
                     elif rule.get("status") == "active":
-                        container_changed = rule.get("container_id") != running_details["container_id"]; service_changed = rule.get("service") != running_details["service"]
-                        if container_changed: logging.info(f"[Reconcile] Updating container ID for active rule {hostname}."); rule["container_id"] = running_details["container_id"]; state_changed_locally = True
-                        if service_changed: logging.info(f"[Reconcile] Updating service for active rule {hostname}."); rule["service"] = running_details["service"]; state_changed_locally = True; needs_cf_update = True
-                        if zone_id_changed: logging.warning(f"[Reconcile] Zone ID for active rule {hostname} changed ('{rule.get('zone_id')}' -> '{target_zone_id}'). Updating state."); rule["zone_id"] = target_zone_id; state_changed_locally = True; hostnames_requiring_dns_check.append(hostname); needs_cf_update = True
-                else: logging.info(f"[Reconcile] Found running container for new hostname {hostname}. Adding rule."); managed_rules[hostname] = {"service": running_details["service"], "container_id": running_details["container_id"], "status": "active", "delete_at": None, "zone_id": target_zone_id}; state_changed_locally = True; needs_cf_update = True; hostnames_requiring_dns_check.append(hostname)
-            for hostname in list(managed_hostnames):
-                if hostname not in running_hostnames and hostname in managed_rules and managed_rules[hostname].get("status") == "active": logging.info(f"[Reconcile] Rule {hostname} is active in state but no matching container running. Scheduling deletion."); rule = managed_rules[hostname]; rule["status"] = "pending_deletion"; rule["delete_at"] = now_utc + timedelta(seconds=GRACE_PERIOD_SECONDS); state_changed_locally = True
-            logging.debug("[Reconcile] Fetching current CF config for final comparison..."); current_cf_config = get_current_cf_config()
+                        # Rule active, check for changes in service, container ID, or zone
+                        container_changed = rule.get("container_id") != running_details["container_id"]
+                        service_changed = rule.get("service") != running_details["service"]
+                        if container_changed:
+                             logging.info(f"[Reconcile] Updating container ID for active rule {hostname}.");
+                             rule["container_id"] = running_details["container_id"]; state_changed_locally = True
+                        if service_changed:
+                             logging.info(f"[Reconcile] Updating service for active rule {hostname}.");
+                             rule["service"] = running_details["service"]; state_changed_locally = True; needs_cf_update = True
+                        if zone_id_changed:
+                             logging.warning(f"[Reconcile] Zone ID for active rule {hostname} changed ('{rule.get('zone_id')}' -> '{target_zone_id}'). Updating state.");
+                             rule["zone_id"] = target_zone_id; state_changed_locally = True;
+                             # DNS needs checking in the new zone
+                             hostnames_requiring_dns_check.append(hostname)
+                             # CF config update might be needed if service also changed, or just for consistency
+                             needs_cf_update = True
+                else:
+                    # New rule for a running container not previously managed
+                    logging.info(f"[Reconcile] Found running container for new hostname {hostname}. Adding rule.")
+                    managed_rules[hostname] = {
+                        "service": running_details["service"], "container_id": running_details["container_id"],
+                        "status": "active", "delete_at": None, "zone_id": target_zone_id
+                    }
+                    state_changed_locally = True; needs_cf_update = True
+                    hostnames_requiring_dns_check.append(hostname) # Ensure DNS exists
+
+            # 2. Check managed rules: if active rule's container is NOT running, schedule deletion
+            for hostname in list(managed_hostnames): # Iterate over copy as we might modify dict
+                if hostname not in running_hostnames:
+                     # Check rule still exists and is active before scheduling deletion
+                     if hostname in managed_rules and managed_rules[hostname].get("status") == "active":
+                         logging.info(f"[Reconcile] Rule {hostname} is active in state but no matching container running. Scheduling deletion.")
+                         rule = managed_rules[hostname]
+                         rule["status"] = "pending_deletion"
+                         rule["delete_at"] = now_utc + timedelta(seconds=GRACE_PERIOD_SECONDS)
+                         state_changed_locally = True
+
+            # 3. Compare Local State (active rules) with Cloudflare State (tunnel ingress)
+            # This check helps catch drift if CF was modified manually or an update failed previously.
+            logging.debug("[Reconcile] Fetching current CF config for final comparison...")
+            current_cf_config = get_current_cf_config()
             if current_cf_config is not None:
                 cf_ingress_hostnames = {r.get("hostname") for r in current_cf_config.get("ingress", []) if r.get("hostname") and r.get("service") != "http_status:404"}
                 active_managed_hostnames = {hn for hn, d in managed_rules.items() if d.get("status") == "active"}
-                if cf_ingress_hostnames != active_managed_hostnames: logging.warning(f"[Reconcile] Mismatch detected between locally managed active rules ({len(active_managed_hostnames)}) and Cloudflare tunnel config ({len(cf_ingress_hostnames)})!"); logging.info(f"[Reconcile] Hostnames managed locally (active): {sorted(list(active_managed_hostnames))}"); logging.info(f"[Reconcile] Hostnames found in Cloudflare config: {sorted(list(cf_ingress_hostnames))}"); needs_cf_update = True
-                else: logging.debug("[Reconcile] Locally managed active rules match Cloudflare tunnel config.")
-            else: logging.error("[Reconcile] Could not fetch current Cloudflare config for comparison.")
+
+                if cf_ingress_hostnames != active_managed_hostnames:
+                     logging.warning(f"[Reconcile] Mismatch detected between locally managed active rules ({len(active_managed_hostnames)}) and Cloudflare tunnel config ({len(cf_ingress_hostnames)})!")
+                     logging.info(f"[Reconcile] Hostnames managed locally (active): {sorted(list(active_managed_hostnames))}")
+                     logging.info(f"[Reconcile] Hostnames found in Cloudflare config: {sorted(list(cf_ingress_hostnames))}")
+                     # Mark for update to push local state to Cloudflare
+                     needs_cf_update = True
+                else:
+                     logging.debug("[Reconcile] Locally managed active rules match Cloudflare tunnel config.")
+            else:
+                 # Cannot compare if fetch failed
+                 logging.error("[Reconcile] Could not fetch current Cloudflare config for comparison.")
+
+            # Save state if any local changes were made
             if state_changed_locally:
                 logging.info("[Reconcile] Saving local state changes made during reconciliation.")
                 save_state()
-                send_sse_event('rules_update') # Signal after state potentially changed
+
             logging.debug("[Reconcile] Releasing state lock.")
+        # --- End state lock ---
+
+        # --- Trigger Updates if needed ---
         if needs_cf_update:
             logging.info("[Reconcile] Triggering Cloudflare tunnel config update based on reconciliation results.")
             if update_cloudflare_config():
-                 if hostnames_requiring_dns_check: logging.info(f"[Reconcile] Checking/Creating DNS records for new/reactivated rules: {hostnames_requiring_dns_check}")
-                 for hostname in hostnames_requiring_dns_check:
-                     rule_details = None; with state_lock: rule_details = managed_rules.get(hostname)
-                     if rule_details and rule_details.get("zone_id") and tunnel_state.get("id"):
-                         if not create_cloudflare_dns_record(rule_details["zone_id"], hostname, tunnel_state["id"]): logging.error(f"[Reconcile] CRITICAL: Failed DNS check/create for {hostname} in zone {rule_details['zone_id']} after tunnel config update.")
-                         else: logging.debug(f"[Reconcile] DNS check/create successful for {hostname} in zone {rule_details['zone_id']}.")
-                     else: logging.error(f"[Reconcile] Cannot check/create DNS for {hostname}: Missing rule details, zone ID, or tunnel ID in current state.")
-            else: logging.error("[Reconcile] Failed Cloudflare tunnel config update during reconciliation. DNS checks for new/reactivated rules skipped.")
-        elif state_changed_locally: logging.info("[Reconcile] Local state changes made (e.g., scheduling deletion), no immediate Cloudflare config update required.")
-        else: logging.info("[Reconcile] No changes required based on reconciliation.")
-    except Exception as e: logging.error(f"Unexpected error during state reconciliation: {e}", exc_info=True)
-    finally: logging.info("Reconciliation complete.")
+                 # If tunnel config update succeeded, ensure DNS records are correct for affected hostnames
+                 if hostnames_requiring_dns_check:
+                      logging.info(f"[Reconcile] Checking/Creating DNS records for new/reactivated rules: {hostnames_requiring_dns_check}")
+                      for hostname in hostnames_requiring_dns_check:
+                           rule_details = None
+                           with state_lock: # Get rule details safely again
+                               rule_details = managed_rules.get(hostname)
+                           if rule_details and rule_details.get("zone_id") and tunnel_state.get("id"):
+                                # Attempt to create/verify DNS record
+                                if not create_cloudflare_dns_record(rule_details["zone_id"], hostname, tunnel_state["id"]):
+                                     logging.error(f"[Reconcile] CRITICAL: Failed DNS check/create for {hostname} in zone {rule_details['zone_id']} after tunnel config update.")
+                                else:
+                                     logging.debug(f"[Reconcile] DNS check/create successful for {hostname} in zone {rule_details['zone_id']}.")
+                           else:
+                                logging.error(f"[Reconcile] Cannot check/create DNS for {hostname}: Missing rule details, zone ID, or tunnel ID in current state.")
+            else:
+                 logging.error("[Reconcile] Failed Cloudflare tunnel config update during reconciliation. DNS checks for new/reactivated rules skipped.")
+        elif state_changed_locally:
+            # e.g., only rules scheduled for deletion, no immediate CF update needed
+            logging.info("[Reconcile] Local state changes made (e.g., scheduling deletion), no immediate Cloudflare config update required.")
+        else:
+            logging.info("[Reconcile] No changes required based on reconciliation.")
+
+    except Exception as e:
+        logging.error(f"Unexpected error during state reconciliation: {e}", exc_info=True)
+    finally:
+        logging.info("Reconciliation complete.")
+
 
 def get_cloudflared_container():
-    if not docker_client: logging.warning("Docker client unavailable."); return None
-    try: return docker_client.containers.get(CLOUDFLARED_CONTAINER_NAME)
-    except NotFound: logging.debug(f"Container '{CLOUDFLARED_CONTAINER_NAME}' not found."); return None
-    except APIError as e: logging.error(f"Docker API error getting container '{CLOUDFLARED_CONTAINER_NAME}': {e}"); cloudflared_agent_state["last_action_status"] = f"Error get agent: {e}"; return None
-    except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection error getting container: {e}"); cloudflared_agent_state["last_action_status"] = f"Error connect Docker: {e}"; return None
-    except Exception as e: logging.error(f"Unexpected error getting container '{CLOUDFLARED_CONTAINER_NAME}': {e}", exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error unexpected get agent: {e}"; return None
+    """Gets the cloudflared agent container object."""
+    if not docker_client:
+        logging.warning("Docker client unavailable.")
+        return None
+    try:
+        return docker_client.containers.get(CLOUDFLARED_CONTAINER_NAME)
+    except NotFound:
+        logging.debug(f"Container '{CLOUDFLARED_CONTAINER_NAME}' not found.")
+        return None
+    except APIError as e:
+        logging.error(f"Docker API error getting container '{CLOUDFLARED_CONTAINER_NAME}': {e}")
+        cloudflared_agent_state["last_action_status"] = f"Error get agent: {e}"
+        return None
+    except requests.exceptions.ConnectionError as e:
+        # Often happens if docker daemon is stopped
+        logging.error(f"Docker connection error getting container: {e}")
+        cloudflared_agent_state["last_action_status"] = f"Error connect Docker: {e}"
+        return None
+    except Exception as e:
+        logging.error(f"Unexpected error getting container '{CLOUDFLARED_CONTAINER_NAME}': {e}", exc_info=True)
+        cloudflared_agent_state["last_action_status"] = f"Error unexpected get agent: {e}"
+        return None
+
 
 def update_cloudflared_container_status():
+    """Updates the global state with the agent container's current status."""
     global docker_client
-    if not docker_client: logging.warning("Docker client unavailable, attempting reconnect...");
-    try: docker_client = docker.from_env(timeout=5); docker_client.ping(); logging.info("Reconnected to Docker daemon.")
-    except Exception as e: logging.error(f"Failed to reconnect to Docker daemon: {e}"); cloudflared_agent_state["container_status"] = "docker_unavailable"; docker_client = None; return
+    if not docker_client:
+        logging.warning("Docker client unavailable, attempting reconnect...")
+        try:
+            docker_client = docker.from_env(timeout=5)
+            docker_client.ping()
+            logging.info("Reconnected to Docker daemon.")
+            # If reconnect succeeds, immediately update status
+        except Exception as e:
+            logging.error(f"Failed to reconnect to Docker daemon: {e}")
+            cloudflared_agent_state["container_status"] = "docker_unavailable"
+            docker_client = None # Ensure it's reset
+            return # Cannot update status
+
     container = get_cloudflared_container()
     if container:
-        try: container.reload(); new_status = container.status;
-        if cloudflared_agent_state["container_status"] != new_status: logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' status changed: {cloudflared_agent_state['container_status']} -> {new_status}"); cloudflared_agent_state["container_status"] = new_status; if new_status == 'running': cloudflared_agent_state["last_action_status"] = None
+        try:
+            container.reload() # Get fresh status
+            new_status = container.status
+            if cloudflared_agent_state["container_status"] != new_status:
+                 logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' status changed: {cloudflared_agent_state['container_status']} -> {new_status}")
+                 cloudflared_agent_state["container_status"] = new_status
+                 # Clear last action status if container becomes running
+                 if new_status == 'running':
+                     cloudflared_agent_state["last_action_status"] = None
         except (NotFound, APIError) as e:
-             if cloudflared_agent_state["container_status"] != "not_found": logging.warning(f"Error reloading agent container status (assuming 'not_found'): {e}"); cloudflared_agent_state["container_status"] = "not_found"; cloudflared_agent_state["last_action_status"] = "Agent container disappeared."
-        except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection error updating agent status: {e}"); cloudflared_agent_state["container_status"] = "docker_unavailable"; docker_client = None; return
+             # If reload fails, container is likely gone
+             if cloudflared_agent_state["container_status"] != "not_found":
+                 logging.warning(f"Error reloading agent container status (assuming 'not_found'): {e}")
+                 cloudflared_agent_state["container_status"] = "not_found"
+                 cloudflared_agent_state["last_action_status"] = "Agent container disappeared."
+        except requests.exceptions.ConnectionError as e:
+             # Handle connection error during reload
+             logging.error(f"Docker connection error updating agent status: {e}")
+             cloudflared_agent_state["container_status"] = "docker_unavailable"
+             docker_client = None # Reset client
+             return
     else:
-        if cloudflared_agent_state.get("container_status") not in ["not_found", "docker_unavailable"]: logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' not found."); cloudflared_agent_state["container_status"] = "not_found"
+        # Container not found by get_cloudflared_container
+        if cloudflared_agent_state.get("container_status") not in ["not_found", "docker_unavailable"]:
+            logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' not found.")
+            cloudflared_agent_state["container_status"] = "not_found"
+
 
 def ensure_docker_network_exists(network_name):
-     if not docker_client: logging.error("Docker client unavailable, cannot check/create network."); return False
-     try: docker_client.networks.get(network_name); logging.info(f"Docker network '{network_name}' already exists."); return True
-     except NotFound: logging.info(f"Docker network '{network_name}' not found. Creating...");
-     try: docker_client.networks.create(network_name, driver="bridge", check_duplicate=True); logging.info(f"Successfully created Docker network '{network_name}'."); return True
+     """Checks if the required Docker network exists, creates if not."""
+     if not docker_client:
+         logging.error("Docker client unavailable, cannot check/create network.")
+         return False
+     try:
+         docker_client.networks.get(network_name)
+         logging.info(f"Docker network '{network_name}' already exists.")
+         return True
+     except NotFound:
+        logging.info(f"Docker network '{network_name}' not found. Creating...")
+        try:
+            docker_client.networks.create(network_name, driver="bridge", check_duplicate=True)
+            logging.info(f"Successfully created Docker network '{network_name}'.")
+            return True
+        except APIError as e:
+            # Handle race condition where network is created between get and create
+            if "already exists" in str(e):
+                logging.warning(f"Network '{network_name}' created concurrently? Treating as success.")
+                return True
+            logging.error(f"Failed to create Docker network '{network_name}': {e}", exc_info=True)
+            cloudflared_agent_state["last_action_status"] = f"Error create net: {e}"
+            return False
+        except Exception as e: # Catch other potential errors during creation
+             logging.error(f"Unexpected error creating Docker network '{network_name}': {e}", exc_info=True);
+             cloudflared_agent_state["last_action_status"] = f"Error: Unexpected create net: {e}"; return False
      except APIError as e:
-         if "already exists" in str(e): logging.warning(f"Network '{network_name}' created concurrently? Treating as success."); return True
-         logging.error(f"Failed to create Docker network '{network_name}': {e}", exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error create net: {e}"; return False
-     except Exception as e: logging.error(f"Unexpected error creating Docker network '{network_name}': {e}", exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: Unexpected create net: {e}"; return False
-     except APIError as e: logging.error(f"Docker API error checking network '{network_name}': {e}", exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error check net: {e}"; return False
-     except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection error checking network '{network_name}': {e}"); cloudflared_agent_state["last_action_status"] = f"Error: Docker connect check net."; return False
-     except Exception as e: logging.error(f"Unexpected error checking network '{network_name}': {e}", exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: Unexpected check net: {e}"; return False
+         logging.error(f"Docker API error checking network '{network_name}': {e}", exc_info=True)
+         cloudflared_agent_state["last_action_status"] = f"Error check net: {e}"
+         return False
+     except requests.exceptions.ConnectionError as e:
+         logging.error(f"Docker connection error checking network '{network_name}': {e}")
+         cloudflared_agent_state["last_action_status"] = f"Error: Docker connect check net."
+         return False
+     except Exception as e:
+         logging.error(f"Unexpected error checking network '{network_name}': {e}", exc_info=True)
+         cloudflared_agent_state["last_action_status"] = f"Error: Unexpected check net: {e}"
+         return False
+
 
 def start_cloudflared_container():
-    logging.info(f"Attempting to start agent container '{CLOUDFLARED_CONTAINER_NAME}'..."); cloudflared_agent_state["last_action_status"] = "Starting..."; success_flag = False
+    """Starts or creates the cloudflared agent container."""
+    logging.info(f"Attempting to start agent container '{CLOUDFLARED_CONTAINER_NAME}'...")
+    cloudflared_agent_state["last_action_status"] = "Starting..."
+    success_flag = False
     try:
-        if not docker_client: msg = "Docker client not available."; logging.error(msg); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
-        if not tunnel_state.get("token"): msg = "Tunnel token not available."; logging.error(msg); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
-        if not ensure_docker_network_exists(CLOUDFLARED_NETWORK_NAME): logging.error(f"Failed network check/create for '{CLOUDFLARED_NETWORK_NAME}'. Cannot start agent."); return False
-        token = tunnel_state["token"]; container = get_cloudflared_container(); needs_recreate = False
+        # Pre-checks
+        if not docker_client:
+            msg = "Docker client not available."; logging.error(msg)
+            cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
+        if not tunnel_state.get("token"):
+            msg = "Tunnel token not available."; logging.error(msg)
+            cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
+        if not ensure_docker_network_exists(CLOUDFLARED_NETWORK_NAME):
+             # Error message should be set by ensure_docker_network_exists
+             logging.error(f"Failed network check/create for '{CLOUDFLARED_NETWORK_NAME}'. Cannot start agent.")
+             return False
+
+        token = tunnel_state["token"]
+        container = get_cloudflared_container()
+        needs_recreate = False
+
+        # Check existing container
         if container:
              try:
-                 container.reload(); logging.info(f"Found existing container '{CLOUDFLARED_CONTAINER_NAME}' status: {container.status}")
-                 if container.status == 'running': msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' is already running."; logging.info(msg); cloudflared_agent_state["last_action_status"] = msg; success_flag = True; return True
-                 current_networks = container.attrs.get('NetworkSettings', {}).get('Networks', {}); network_mode = container.attrs.get('HostConfig', {}).get('NetworkMode', 'default')
-                 if network_mode == 'host' or network_mode == CLOUDFLARED_NETWORK_NAME: logging.warning(f"Existing container '{CLOUDFLARED_CONTAINER_NAME}' is in an unexpected network mode ('{network_mode}'). Needs recreation."); needs_recreate = True
-                 elif CLOUDFLARED_NETWORK_NAME not in current_networks: logging.warning(f"Existing container '{CLOUDFLARED_CONTAINER_NAME}' is not attached to the required network '{CLOUDFLARED_NETWORK_NAME}'. Needs recreation."); needs_recreate = True
-                 if needs_recreate: logging.info(f"Removing misconfigured/stopped container '{CLOUDFLARED_CONTAINER_NAME}' before creating a new one...");
-                 try: container.remove(force=True); container = None
-                 except (APIError, requests.exceptions.ConnectionError) as rm_err: logging.error(f"Failed to remove misconfigured container '{CLOUDFLARED_CONTAINER_NAME}': {rm_err}. Cannot proceed."); cloudflared_agent_state["last_action_status"] = f"Error: Failed remove old agent: {rm_err}"; return False
-                 else: logging.info(f"Starting existing stopped container '{CLOUDFLARED_CONTAINER_NAME}'..."); container.start(); msg = f"Started existing container '{CLOUDFLARED_CONTAINER_NAME}'."; cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
-             except (NotFound, APIError) as e: logging.warning(f"Error checking existing container '{CLOUDFLARED_CONTAINER_NAME}': {e}. Assuming creation is needed."); container = None
-             except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection error checking existing container: {e}"); cloudflared_agent_state["last_action_status"] = f"Error: Docker connect check agent."; return False
-        if not container and not success_flag:
+                 container.reload()
+                 logging.info(f"Found existing container '{CLOUDFLARED_CONTAINER_NAME}' status: {container.status}")
+                 if container.status == 'running':
+                     msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' is already running."; logging.info(msg)
+                     cloudflared_agent_state["last_action_status"] = msg; success_flag = True; return True
+
+                 # Check if existing non-running container has correct network config
+                 current_networks = container.attrs.get('NetworkSettings', {}).get('Networks', {})
+                 network_mode = container.attrs.get('HostConfig', {}).get('NetworkMode', 'default')
+                 # If container is in host mode or the specific cloudflare network directly (deprecated pattern)
+                 if network_mode == 'host' or network_mode == CLOUDFLARED_NETWORK_NAME:
+                     logging.warning(f"Existing container '{CLOUDFLARED_CONTAINER_NAME}' is in an unexpected network mode ('{network_mode}'). Needs recreation.")
+                     needs_recreate = True
+                 # If container exists but isn't attached to the target network
+                 elif CLOUDFLARED_NETWORK_NAME not in current_networks:
+                     logging.warning(f"Existing container '{CLOUDFLARED_CONTAINER_NAME}' is not attached to the required network '{CLOUDFLARED_NETWORK_NAME}'. Needs recreation.")
+                     needs_recreate = True
+
+                 if needs_recreate:
+                      logging.info(f"Removing misconfigured/stopped container '{CLOUDFLARED_CONTAINER_NAME}' before creating a new one...")
+                      try:
+                          container.remove(force=True) # Force remove if stopped
+                          container = None # Ensure we proceed to creation logic
+                      except (APIError, requests.exceptions.ConnectionError) as rm_err:
+                           logging.error(f"Failed to remove misconfigured container '{CLOUDFLARED_CONTAINER_NAME}': {rm_err}. Cannot proceed.")
+                           cloudflared_agent_state["last_action_status"] = f"Error: Failed remove old agent: {rm_err}"; return False
+                 else:
+                      # Container exists, is stopped, and seems correctly configured - just start it
+                      logging.info(f"Starting existing stopped container '{CLOUDFLARED_CONTAINER_NAME}'...");
+                      container.start()
+                      msg = f"Started existing container '{CLOUDFLARED_CONTAINER_NAME}'.";
+                      cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
+
+             except (NotFound, APIError) as e:
+                  logging.warning(f"Error checking existing container '{CLOUDFLARED_CONTAINER_NAME}': {e}. Assuming creation is needed.");
+                  container = None # Reset container object
+             except requests.exceptions.ConnectionError as e:
+                  logging.error(f"Docker connection error checking existing container: {e}")
+                  cloudflared_agent_state["last_action_status"] = f"Error: Docker connect check agent."; return False
+
+        # Create container if it doesn't exist or needed recreation
+        if not container and not success_flag: # Only create if not already started/running
             logging.info(f"Container '{CLOUDFLARED_CONTAINER_NAME}' not found or needs recreation. Creating...")
             try:
-                try: logging.info(f"Pulling image {CLOUDFLARED_IMAGE}..."); docker_client.images.pull(CLOUDFLARED_IMAGE); logging.info("Image pull complete.")
-                except APIError as img_err: logging.warning(f"Could not pull image {CLOUDFLARED_IMAGE}: {img_err}. Container run will attempt using local image if available.")
-                except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection failed during image pull: {e}"); cloudflared_agent_state["last_action_status"] = f"Error: Docker connect pull image."; return False
-                container_params = { "image": CLOUDFLARED_IMAGE, "command": f"tunnel --no-autoupdate run --token {token}", "name": CLOUDFLARED_CONTAINER_NAME, "network": CLOUDFLARED_NETWORK_NAME, "restart_policy": {"Name": "unless-stopped"}, "detach": True, "remove": False, "labels": {"managed-by": "dockflare"}}
-                new_container = docker_client.containers.run(**container_params); msg = f"Successfully created and started container '{new_container.name}' ({new_container.id[:12]})."; cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
+                # Pull latest image first
+                try:
+                    logging.info(f"Pulling image {CLOUDFLARED_IMAGE}...");
+                    docker_client.images.pull(CLOUDFLARED_IMAGE);
+                    logging.info("Image pull complete.")
+                except APIError as img_err:
+                    # Log warning but proceed - maybe image exists locally
+                    logging.warning(f"Could not pull image {CLOUDFLARED_IMAGE}: {img_err}. Container run will attempt using local image if available.")
+                except requests.exceptions.ConnectionError as e:
+                    logging.error(f"Docker connection failed during image pull: {e}")
+                    cloudflared_agent_state["last_action_status"] = f"Error: Docker connect pull image."; return False
+
+                # Define container parameters
+                container_params = {
+                    "image": CLOUDFLARED_IMAGE,
+                    "command": f"tunnel --no-autoupdate run --token {token}",
+                    "name": CLOUDFLARED_CONTAINER_NAME,
+                    "network": CLOUDFLARED_NETWORK_NAME, # Attach to custom network
+                    "restart_policy": {"Name": "unless-stopped"},
+                    "detach": True, # Run in background
+                    "remove": False, # Do not auto-remove on exit
+                    "labels": {"managed-by": "dockflare"} # Identify container
+                }
+                # Run the container
+                new_container = docker_client.containers.run(**container_params)
+                msg = f"Successfully created and started container '{new_container.name}' ({new_container.id[:12]})."
+                cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
+
             except APIError as create_err:
-                if "is already in use" in str(create_err): logging.error(f"Container name '{CLOUDFLARED_CONTAINER_NAME}' is already in use by another container."); msg = f"Error: Container name '{CLOUDFLARED_CONTAINER_NAME}' conflict. Remove the conflicting container manually and retry."
-                else: msg = f"Docker API error creating container: {create_err}"; logging.error(msg, exc_info=True)
+                # Handle specific error like name conflict
+                if "is already in use" in str(create_err):
+                    logging.error(f"Container name '{CLOUDFLARED_CONTAINER_NAME}' is already in use by another container.")
+                    msg = f"Error: Container name '{CLOUDFLARED_CONTAINER_NAME}' conflict. Remove the conflicting container manually and retry."
+                else:
+                     msg = f"Docker API error creating container: {create_err}"
+                     logging.error(msg, exc_info=True)
                 cloudflared_agent_state["last_action_status"] = msg; success_flag = False
-            except requests.exceptions.ConnectionError as e: logging.error(f"Docker connection failed running container: {e}"); cloudflared_agent_state["last_action_status"] = f"Error: Docker connect run agent."; success_flag = False
-    except APIError as e: msg = f"Docker API error during start sequence: {e}"; logging.error(msg, exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
-    except requests.exceptions.ConnectionError as e: msg = f"Docker connection error during start sequence: {e}"; logging.error(msg); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
-    except Exception as e: msg = f"Unexpected error starting container: {e}"; logging.error(msg, exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+            except requests.exceptions.ConnectionError as e:
+                logging.error(f"Docker connection failed running container: {e}")
+                cloudflared_agent_state["last_action_status"] = f"Error: Docker connect run agent."; success_flag = False
+
+    # Catch broader errors during the start sequence
+    except APIError as e:
+        msg = f"Docker API error during start sequence: {e}"; logging.error(msg, exc_info=True)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+    except requests.exceptions.ConnectionError as e:
+        msg = f"Docker connection error during start sequence: {e}"; logging.error(msg)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+    except Exception as e:
+        msg = f"Unexpected error starting container: {e}"; logging.error(msg, exc_info=True)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
     finally:
-        if docker_client: logging.debug("Updating agent status after start attempt..."); time.sleep(2); update_cloudflared_container_status()
-        logging.info(f"Exiting start_cloudflared_container function (Success: {success_flag})."); return success_flag
+        # Update status after attempt, slight delay allows container to settle
+        if docker_client:
+             logging.debug("Updating agent status after start attempt...");
+             time.sleep(2);
+             update_cloudflared_container_status()
+        logging.info(f"Exiting start_cloudflared_container function (Success: {success_flag}).")
+        return success_flag
+
 
 def stop_cloudflared_container():
-    logging.info(f"Attempting to stop agent container '{CLOUDFLARED_CONTAINER_NAME}'..."); cloudflared_agent_state["last_action_status"] = "Stopping..."; success_flag = False
+    """Stops the cloudflared agent container."""
+    logging.info(f"Attempting to stop agent container '{CLOUDFLARED_CONTAINER_NAME}'...")
+    cloudflared_agent_state["last_action_status"] = "Stopping..."
+    success_flag = False
     try:
-        if not docker_client: msg = "Docker client unavailable."; logging.error(msg); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
+        if not docker_client:
+            msg = "Docker client unavailable."; logging.error(msg)
+            cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; return False
+
         container = get_cloudflared_container()
-        if not container: msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' not found (already stopped or removed?)."; logging.warning(msg); cloudflared_agent_state["last_action_status"] = msg; if cloudflared_agent_state["container_status"] != "not_found": cloudflared_agent_state["container_status"] = "not_found"; success_flag = True; return True
+        if not container:
+            msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' not found (already stopped or removed?)."; logging.warning(msg)
+            cloudflared_agent_state["last_action_status"] = msg
+            # Ensure status reflects not found if it wasn't already
+            if cloudflared_agent_state["container_status"] != "not_found":
+                 cloudflared_agent_state["container_status"] = "not_found"
+            success_flag = True; return True # Treat as success if already gone
+
+        # Check status before stopping
         container.reload()
-        if container.status != 'running': msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' is not running (status: {container.status})."; logging.info(msg); cloudflared_agent_state["last_action_status"] = msg; if cloudflared_agent_state["container_status"] != container.status: cloudflared_agent_state["container_status"] = container.status; success_flag = True; return True
-        logging.info(f"Stopping running container '{CLOUDFLARED_CONTAINER_NAME}'..."); container.stop(timeout=30); msg = f"Successfully stopped container '{CLOUDFLARED_CONTAINER_NAME}'."; cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
-    except (APIError, NotFound) as e: msg = f"Docker API error stopping container '{CLOUDFLARED_CONTAINER_NAME}': {e}"; logging.error(msg, exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
-    except requests.exceptions.ConnectionError as e: msg = f"Docker connection error stopping container: {e}"; logging.error(msg); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
-    except Exception as e: msg = f"Unexpected error stopping container '{CLOUDFLARED_CONTAINER_NAME}': {e}"; logging.error(msg, exc_info=True); cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+        if container.status != 'running':
+            msg = f"Container '{CLOUDFLARED_CONTAINER_NAME}' is not running (status: {container.status})."; logging.info(msg)
+            cloudflared_agent_state["last_action_status"] = msg
+            # Update status if needed
+            if cloudflared_agent_state["container_status"] != container.status:
+                 cloudflared_agent_state["container_status"] = container.status
+            success_flag = True; return True # Treat as success if not running
+
+        # Stop the running container
+        logging.info(f"Stopping running container '{CLOUDFLARED_CONTAINER_NAME}'...");
+        container.stop(timeout=30) # Allow 30 seconds for graceful shutdown
+        msg = f"Successfully stopped container '{CLOUDFLARED_CONTAINER_NAME}'.";
+        cloudflared_agent_state["last_action_status"] = msg; logging.info(msg); success_flag = True
+
+    except (APIError, NotFound) as e:
+        # NotFound might occur if container removed between get and stop
+        msg = f"Docker API error stopping container '{CLOUDFLARED_CONTAINER_NAME}': {e}"; logging.error(msg, exc_info=True)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+    except requests.exceptions.ConnectionError as e:
+        msg = f"Docker connection error stopping container: {e}"; logging.error(msg)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
+    except Exception as e:
+        msg = f"Unexpected error stopping container '{CLOUDFLARED_CONTAINER_NAME}': {e}"; logging.error(msg, exc_info=True)
+        cloudflared_agent_state["last_action_status"] = f"Error: {msg}"; success_flag = False
     finally:
-        if docker_client: logging.debug("Updating agent status after stop attempt..."); time.sleep(2); update_cloudflared_container_status()
-        logging.info(f"Exiting stop_cloudflared_container function (Success: {success_flag})."); return success_flag
+        # Update status after attempt
+        if docker_client:
+             logging.debug("Updating agent status after stop attempt...");
+             time.sleep(2); # Give docker time to update status
+             update_cloudflared_container_status()
+        logging.info(f"Exiting stop_cloudflared_container function (Success: {success_flag}).")
+        return success_flag
+
 
 # --- Flask App Setup ---
 app = Flask(__name__)
-app.secret_key = os.urandom(24)
+app.secret_key = os.urandom(24) # For potential future session use
+
 
 def get_display_token(token):
-    if not token: return "Not available"
+    """Returns a truncated token for display."""
+    if not token:
+        return "Not available"
     return f"{token[:5]}...{token[-5:]}" if len(token) > 10 else "Token retrieved (short)"
+
 
 # --- Flask Routes ---
 @app.route('/')
 def status_page():
-    update_cloudflared_container_status()
+    """Renders the main status dashboard page."""
+    update_cloudflared_container_status() # Get latest status before rendering
     with state_lock:
+        # Create a deep copy for template rendering to avoid race conditions
+        # and handle datetime serialization for Jinja
         rules_for_template = json.loads(json.dumps(managed_rules, default=str))
+        # Convert delete_at back to datetime if needed for template logic (like countdown)
         for rule in rules_for_template.values():
              if rule.get("delete_at") and isinstance(rule.get("delete_at"), str):
                  try:
-                     if rule["delete_at"].endswith('Z'): dt = datetime.fromisoformat(rule["delete_at"].replace('Z', '+00:00'))
-                     else: dt = datetime.fromisoformat(rule["delete_at"])
+                     if rule["delete_at"].endswith('Z'):
+                        dt = datetime.fromisoformat(rule["delete_at"].replace('Z', '+00:00'))
+                     else:
+                         dt = datetime.fromisoformat(rule["delete_at"])
+                     # Ensure timezone aware
                      rule["delete_at"] = dt.replace(tzinfo=timezone.utc) if dt.tzinfo is None else dt.astimezone(timezone.utc)
-                 except Exception as date_parse_err: logging.warning(f"Error parsing delete_at ('{rule['delete_at']}') for template: {date_parse_err}"); rule["delete_at"] = None
-        template_tunnel_state = tunnel_state.copy(); template_agent_state = cloudflared_agent_state.copy()
-    display_token = get_display_token(template_tunnel_state.get("token")); docker_available = docker_client is not None
-    return render_template('status_page.html', tunnel_state=template_tunnel_state, agent_state=template_agent_state, display_token=display_token, cloudflared_container_name=CLOUDFLARED_CONTAINER_NAME, docker_available=docker_available, rules=rules_for_template)
+                 except Exception as date_parse_err:
+                      logging.warning(f"Error parsing delete_at ('{rule['delete_at']}') for template: {date_parse_err}")
+                      rule["delete_at"] = None # Reset if parsing fails
 
-@app.route('/start-tunnel', methods=['POST'])
+        template_tunnel_state = tunnel_state.copy()
+        template_agent_state = cloudflared_agent_state.copy()
+
+    display_token = get_display_token(template_tunnel_state.get("token"))
+    docker_available = docker_client is not None
+
+    return render_template('status_page.html',
+                            tunnel_state=template_tunnel_state,
+                            agent_state=template_agent_state,
+                            display_token=display_token,
+                            cloudflared_container_name=CLOUDFLARED_CONTAINER_NAME,
+                            docker_available=docker_available,
+                            rules=rules_for_template)
+
+
+@app.route('/start-tunnel', methods=['POST']) # Ensure URL matches url_for('start_tunnel')
 def start_tunnel():
-    # Return simple success/failure for JS fetch call
+    """Handles request to start the tunnel agent."""
     logging.info("UI request: Start tunnel agent.")
-    success = start_cloudflared_container()
-    time.sleep(1) # Allow status to potentially update
-    return jsonify(success=success)
+    start_cloudflared_container()
+    time.sleep(1) # Short delay to allow status update
+    return redirect(url_for('status_page'))
 
-@app.route('/stop-tunnel', methods=['POST'])
+
+@app.route('/stop-tunnel', methods=['POST']) # Ensure URL matches url_for('stop_tunnel')
 def stop_tunnel():
-    # Return simple success/failure for JS fetch call
+    """Handles request to stop the tunnel agent."""
     logging.info("UI request: Stop tunnel agent.")
-    success = stop_cloudflared_container()
-    time.sleep(1) # Allow status to potentially update
-    return jsonify(success=success)
+    stop_cloudflared_container()
+    time.sleep(1) # Short delay to allow status update
+    return redirect(url_for('status_page'))
 
-@app.route('/force_delete_rule/<hostname>', methods=['POST'])
+
+@app.route('/force_delete_rule/<hostname>', methods=['POST']) # Ensure URL matches url_for('force_delete_rule')
 def force_delete_rule(hostname):
-    # Keep redirect for this as it's a standard form submission
-    logging.info(f"UI request: Force delete rule for hostname: {hostname}"); rule_removed_from_state = False; dns_delete_success = False; zone_id_for_delete = None
+    """Handles request to immediately delete a rule and its DNS record."""
+    logging.info(f"UI request: Force delete rule for hostname: {hostname}")
+    rule_removed_from_state = False
+    dns_delete_success = False
+    zone_id_for_delete = None
+
+    # Step 0: Get the zone ID from the state BEFORE deleting the rule from state
     with state_lock:
         rule_details = managed_rules.get(hostname)
-        if rule_details: zone_id_for_delete = rule_details.get("zone_id"); else: logging.warning(f"Rule {hostname} not found in state during force delete. Attempting DNS delete in default zone ID ({CF_ZONE_ID}) if available."); zone_id_for_delete = CF_ZONE_ID
+        if rule_details:
+            zone_id_for_delete = rule_details.get("zone_id")
+        else:
+            # Rule already gone from state, maybe try default?
+            logging.warning(f"Rule {hostname} not found in state during force delete. Attempting DNS delete in default zone ID ({CF_ZONE_ID}) if available.")
+            zone_id_for_delete = CF_ZONE_ID # Best effort
+
+    # Step 1: Delete DNS record immediately (if we have a zone ID and tunnel ID)
     if zone_id_for_delete and tunnel_state.get("id"):
-        logging.info(f"Attempting immediate DNS record deletion for force-deleted rule: {hostname} in zone {zone_id_for_delete}"); dns_delete_success = delete_cloudflare_dns_record(zone_id_for_delete, hostname, tunnel_state["id"])
-        if not dns_delete_success: logging.error(f"Failed immediate DNS delete for {hostname} in zone {zone_id_for_delete}. Tunnel config update will proceed."); cloudflared_agent_state["last_action_status"] = f"Warning: Failed DNS delete for {hostname}. Tunnel update proceeding."
-    elif not zone_id_for_delete: logging.error(f"Cannot delete DNS for {hostname}: Zone ID could not be determined."); cloudflared_agent_state["last_action_status"] = f"Error: Cannot delete DNS for {hostname} (missing zone ID)."
-    else: logging.error(f"Cannot delete DNS for {hostname}: Missing Tunnel ID."); cloudflared_agent_state["last_action_status"] = f"Error: Cannot delete DNS for {hostname} (missing tunnel ID)."
+        logging.info(f"Attempting immediate DNS record deletion for force-deleted rule: {hostname} in zone {zone_id_for_delete}")
+        dns_delete_success = delete_cloudflare_dns_record(zone_id_for_delete, hostname, tunnel_state["id"])
+        if not dns_delete_success:
+             logging.error(f"Failed immediate DNS delete for {hostname} in zone {zone_id_for_delete}. Tunnel config update will proceed.")
+             cloudflared_agent_state["last_action_status"] = f"Warning: Failed DNS delete for {hostname}. Tunnel update proceeding." # Inform user
+    elif not zone_id_for_delete:
+        logging.error(f"Cannot delete DNS for {hostname}: Zone ID could not be determined.")
+        cloudflared_agent_state["last_action_status"] = f"Error: Cannot delete DNS for {hostname} (missing zone ID)."
+    else: # Missing tunnel ID
+        logging.error(f"Cannot delete DNS for {hostname}: Missing Tunnel ID.")
+        cloudflared_agent_state["last_action_status"] = f"Error: Cannot delete DNS for {hostname} (missing tunnel ID)."
+
+    # Step 2: Remove rule from local state (under lock)
     with state_lock:
-        if hostname in managed_rules: logging.info(f"Force deleting rule for {hostname} from local state."); del managed_rules[hostname]; rule_removed_from_state = True; save_state(); send_sse_event('rules_update')
-        else: logging.warning(f"Rule '{hostname}' was already removed from state when force delete requested."); rule_removed_from_state = True
+        if hostname in managed_rules:
+            logging.info(f"Force deleting rule for {hostname} from local state.")
+            del managed_rules[hostname]
+            rule_removed_from_state = True
+            save_state() # Save state immediately after removal
+        else:
+            logging.warning(f"Rule '{hostname}' was already removed from state when force delete requested.")
+            rule_removed_from_state = True # Consider it removed for CF update trigger
+
+    # Step 3: Trigger Cloudflare tunnel config update (outside lock)
     if rule_removed_from_state:
         logging.info(f"Triggering Cloudflare tunnel config update after force deleting {hostname}.")
-        if update_cloudflare_config(): logging.info(f"CF tunnel config update successful after force deleting {hostname}."); status_msg = f"Successfully force deleted rule for {hostname} and updated Cloudflare."; status_msg += " (Note: DNS deletion failed or was skipped)." if not dns_delete_success else ""; cloudflared_agent_state["last_action_status"] = status_msg
-        else: logging.error(f"CRITICAL: State updated after force delete of {hostname} (DNS delete success: {dns_delete_success}), but subsequent tunnel config update FAILED!"); cloudflared_agent_state["last_action_status"] = f"Error: Removed {hostname} locally (DNS delete: {dns_delete_success}), but FAILED tunnel config update! Reconciliation needed."
-    time.sleep(1); return redirect(url_for('status_page'))
+        if update_cloudflare_config():
+            logging.info(f"CF tunnel config update successful after force deleting {hostname}.")
+            status_msg = f"Successfully force deleted rule for {hostname} and updated Cloudflare."
+            if not dns_delete_success:
+                status_msg += " (Note: DNS deletion failed or was skipped)."
+            cloudflared_agent_state["last_action_status"] = status_msg
+        else:
+            # This is problematic: state is changed, but CF config failed. Reconciliation needed.
+            logging.error(f"CRITICAL: State updated after force delete of {hostname} (DNS delete success: {dns_delete_success}), but subsequent tunnel config update FAILED!")
+            cloudflared_agent_state["last_action_status"] = f"Error: Removed {hostname} locally (DNS delete: {dns_delete_success}), but FAILED tunnel config update! Reconciliation needed."
 
-@app.route('/get_rules_table')
-def get_rules_table():
-    with state_lock:
-        rules_for_template = json.loads(json.dumps(managed_rules, default=str))
-        for rule in rules_for_template.values():
-             if rule.get("delete_at") and isinstance(rule.get("delete_at"), str):
-                 try:
-                     if rule["delete_at"].endswith('Z'): dt = datetime.fromisoformat(rule["delete_at"].replace('Z', '+00:00'))
-                     else: dt = datetime.fromisoformat(rule["delete_at"])
-                     rule["delete_at"] = dt.replace(tzinfo=timezone.utc) if dt.tzinfo is None else dt.astimezone(timezone.utc)
-                 except Exception: rule["delete_at"] = None
-    docker_available = docker_client is not None
-    return render_template('rules_table_partial.html', rules=rules_for_template, docker_available=docker_available)
+    time.sleep(1) # Short delay before redirect
+    return redirect(url_for('status_page'))
+
 
 @app.route('/stream-logs')
 def stream_logs():
+    """Streams log messages using Server-Sent Events."""
     @stream_with_context
     def event_stream():
-        logging.info("SSE client connected.")
-        yield f"event: connection_status\ndata: {json.dumps({'status': 'connected'})}\n\n"
+        logging.info("Log stream client connected.")
+        # Send a connection confirmation message
+        yield f"data: --- Log stream connected ---\n\n"
         try:
-            while True: sse_message = log_queue.get(block=True); yield sse_message
-        except GeneratorExit: logging.info("SSE client disconnected.")
-        finally: pass
+            while True:
+                # Block waiting for a log message from the queue
+                log_entry = log_queue.get(block=True)
+                # Format for SSE: "data: <message>\n\n"
+                yield f"data: {log_entry}\n\n"
+        except GeneratorExit:
+            # This happens when the client disconnects
+            logging.info("Log stream client disconnected.")
+        finally:
+            # No specific cleanup needed for queue handler in this setup
+            pass
+    # Return the streaming response
     return Response(event_stream(), mimetype='text/event-stream')
+
 
 # --- Background Task Runner ---
 def run_background_tasks():
-    if not docker_client or not tunnel_state.get("id"): logging.warning("Docker or Tunnel not ready. Background tasks won't start."); return None, None
-    logging.info("Starting background threads (Docker Listener, Cleanup Task)..."); event_thread = threading.Thread(target=docker_event_listener, name="DockerEventListener", daemon=True); cleanup_thread = threading.Thread(target=cleanup_expired_rules, name="CleanupTask", daemon=True); event_thread.start(); cleanup_thread.start()
-    logging.info("Background threads started."); return event_thread, cleanup_thread
+    """Starts the Docker event listener and cleanup threads."""
+    if not docker_client:
+        logging.warning("Docker client unavailable. Background tasks cannot start.")
+        return None, None
+    if not tunnel_state.get("id"):
+        logging.warning("Tunnel not initialized. Background tasks cannot start.")
+        return None, None
+
+    logging.info("Starting background threads (Docker Listener, Cleanup Task)...");
+    event_thread = threading.Thread(target=docker_event_listener, name="DockerEventListener", daemon=True)
+    cleanup_thread = threading.Thread(target=cleanup_expired_rules, name="CleanupTask", daemon=True)
+    event_thread.start()
+    cleanup_thread.start()
+    logging.info("Background threads started.")
+    return event_thread, cleanup_thread
+
 
 # --- Main Execution ---
 if __name__ == '__main__':
-    logging.info("-" * 52); logging.info("--- Dockflare Starting ---"); logging.info("-" * 52)
-    load_state(); logging.info("Initial state loading complete.")
-    event_thread = None; cleanup_thread = None
-    if not docker_client: logging.error("Docker client unavailable at startup. Dockflare will run with limited functionality."); tunnel_state["status_message"] = "Error: Docker client unavailable."; tunnel_state["error"] = "Failed to connect to Docker daemon."; cloudflared_agent_state["container_status"] = "docker_unavailable"; logging.warning("Skipping tunnel initialization, reconciliation, agent management, and background tasks due to Docker connection failure.")
+    logging.info("-" * 52)
+    logging.info("--- Dockflare Starting ---")
+    logging.info("-" * 52)
+
+    load_state()
+    logging.info("Initial state loading complete.")
+    event_thread = None
+    cleanup_thread = None
+
+    # --- Critical Pre-checks & Initial Setup ---
+    if not docker_client:
+         logging.error("Docker client unavailable at startup. Dockflare will run with limited functionality.")
+         tunnel_state["status_message"] = "Error: Docker client unavailable."
+         tunnel_state["error"] = "Failed to connect to Docker daemon."
+         cloudflared_agent_state["container_status"] = "docker_unavailable"
+         logging.warning("Skipping tunnel initialization, reconciliation, agent management, and background tasks due to Docker connection failure.")
     else:
+         # --- Normal Startup Flow (Docker Available) ---
          logging.info("Docker client available.")
-         initialize_tunnel(); logging.info(f"Tunnel initialization complete. Status: {tunnel_state.get('status_message')}")
+
+         # Initialize Tunnel (find or create)
+         initialize_tunnel()
+         logging.info(f"Tunnel initialization complete. Status: {tunnel_state.get('status_message')}")
+
+         # Proceed only if tunnel is properly initialized
          if tunnel_state.get("id") and tunnel_state.get("token"):
-             logging.info("Tunnel initialized. Proceeding with initial reconciliation & agent checks."); reconcile_state(); logging.info("Initial state reconciliation complete.")
-             logging.info("Checking cloudflared agent container status..."); update_cloudflared_container_status()
-             if cloudflared_agent_state.get("container_status") != 'running': logging.info("Agent container not running, attempting auto-start..."); start_cloudflared_container()
-             else: logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' is already running.")
+             logging.info("Tunnel initialized. Proceeding with initial reconciliation & agent checks.")
+
+             # Perform initial state reconciliation
+             reconcile_state()
+             logging.info("Initial state reconciliation complete.")
+
+             # Check agent status and start if needed
+             logging.info("Checking cloudflared agent container status...")
+             update_cloudflared_container_status()
+             if cloudflared_agent_state.get("container_status") != 'running':
+                 logging.info("Agent container not running, attempting auto-start...")
+                 start_cloudflared_container()
+             else:
+                 logging.info(f"Agent container '{CLOUDFLARED_CONTAINER_NAME}' is already running.")
+
+             # Start background tasks only if Docker and Tunnel are ready
              event_thread, cleanup_thread = run_background_tasks()
-         else: logging.warning("Tunnel not fully initialized (missing ID or Token). Skipping reconciliation, agent start, and background tasks."); if not tunnel_state.get("error"): tunnel_state["status_message"] = "Tunnel setup incomplete (missing ID/Token)."
+         else:
+             # Tunnel initialization failed
+             logging.warning("Tunnel not fully initialized (missing ID or Token). Skipping reconciliation, agent start, and background tasks.")
+             if not tunnel_state.get("error"): # Ensure a message is set if no specific error was caught
+                 tunnel_state["status_message"] = "Tunnel setup incomplete (missing ID/Token)."
+
+    # --- Start Web Server ---
     logging.info("Starting Flask web server...")
-    flask_thread = None
+    flask_thread = None # Keep track of the server thread
     try:
+        # Use Waitress if available (recommended for production)
         from waitress import serve
-        flask_thread = threading.Thread(target=serve, args=(app,), kwargs={'host':'0.0.0.0','port':5000}, daemon=True, name="FlaskWaitressServer"); flask_thread.start(); logging.info("Flask server started using waitress on 0.0.0.0:5000.")
+        # Run waitress in a separate thread so main thread can monitor
+        flask_thread = threading.Thread(
+            target=serve,
+            args=(app,),
+            kwargs={'host':'0.0.0.0','port':5000},
+            daemon=True,
+            name="FlaskWaitressServer"
+        )
+        flask_thread.start()
+        logging.info("Flask server started using waitress on 0.0.0.0:5000.")
+
+        # Keep main thread alive to monitor background threads (optional but good practice)
         while True:
              all_threads_alive = True
-             if flask_thread and not flask_thread.is_alive(): logging.error("Flask server thread terminated unexpectedly!"); all_threads_alive = False
-             if event_thread and not event_thread.is_alive(): logging.warning("Docker listener thread terminated.")
-             if cleanup_thread and not cleanup_thread.is_alive(): logging.warning("Cleanup thread terminated.")
-             if not all_threads_alive: logging.error("A critical thread terminated. Initiating shutdown."); stop_event.set(); break
-             if stop_event.is_set(): logging.info("Stop event detected by main thread."); break
-             time.sleep(10)
+             if flask_thread and not flask_thread.is_alive():
+                 logging.error("Flask server thread terminated unexpectedly!")
+                 all_threads_alive = False
+             # Check background threads only if they were started
+             if event_thread and not event_thread.is_alive():
+                 logging.warning("Docker listener thread terminated.")
+                 # Decide if this is critical - maybe try restarting it? For now, just log.
+             if cleanup_thread and not cleanup_thread.is_alive():
+                 logging.warning("Cleanup thread terminated.")
+                 # Decide if critical. For now, just log.
+
+             if not all_threads_alive:
+                 logging.error("A critical thread terminated. Initiating shutdown.")
+                 stop_event.set() # Signal other threads to stop
+                 break
+             if stop_event.is_set(): # Allow external signal to stop
+                 logging.info("Stop event detected by main thread.")
+                 break
+             time.sleep(10) # Check periodically
+
     except ImportError:
+        # Fallback to Flask development server if Waitress isn't installed
         logging.warning("Waitress not found. Using Flask development server (not recommended for production). Install using: pip install waitress")
-        app.run(host='0.0.0.0', port=5000, threaded=True, debug=False)
-    except KeyboardInterrupt: logging.info("KeyboardInterrupt received. Shutting down...")
-    except Exception as server_err: logging.error(f"Web server failed unexpectedly: {server_err}", exc_info=True)
+        # Run dev server directly (blocks main thread) - Use threaded for SSE
+        app.run(host='0.0.0.0', port=5000, threaded=True, debug=False) # Debug off for basic logging
+    except KeyboardInterrupt:
+        logging.info("KeyboardInterrupt received. Shutting down...")
+    except Exception as server_err:
+        # Catch other potential server errors
+        logging.error(f"Web server failed unexpectedly: {server_err}", exc_info=True)
     finally:
-        logging.info("Shutdown initiated..."); stop_event.set(); logging.info("Stop event set for background tasks.")
+        # --- Graceful Shutdown ---
+        logging.info("Shutdown initiated...")
+        stop_event.set() # Signal background threads to stop gracefully
+        logging.info("Stop event set for background tasks.")
+
+        # Optionally wait for threads to finish (add timeout)
+        # if event_thread and event_thread.is_alive(): event_thread.join(timeout=5)
+        # if cleanup_thread and cleanup_thread.is_alive(): cleanup_thread.join(timeout=5)
+        # Note: Flask server thread (if run directly) or Waitress thread might need separate handling if not daemon
+
         logging.info("Exiting Dockflare application.")
+        # Exit with error code if critical components failed
         exit_code = 1 if tunnel_state.get("error") or cloudflared_agent_state.get("container_status") == "docker_unavailable" else 0
         sys.exit(exit_code)
